### 2025-11-05 #9352 @editorialbot
[PRE REVIEW]: Liblsl.dart: A Dart native API for Lab Streaming Layer (LSL)
**Submitting author:** @zeyus (<a href="http://orcid.org/0009-0000-0930-4172">Luke Daniel Ring</a>)
**Repository:** https://github.com/NexusDynamic/liblsl.dart
**Branch with paper.md** (empty if default branch):
**Version:** v0.9.1
