### 2025-11-05 18f4f90
Bump version, update liblsl ref (and submodule dir).
