+++ packages/liblsl/pubspec.yaml
@@ -1,6 +1,6 @@
 name: liblsl
 description: A Dart (and Flutter) native library for working with Lab Streaming Layer (LSL / liblsl).
-version: 0.10.0+1
+version: 0.10.1+0
 homepage: https://zeyus.com/
 repository: https://github.com/NexusDynamic/liblsl.dart
 issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
@@ -33,12 +33,12 @@ ffigen:
   output: 'lib/native_liblsl.dart'
   headers:
     entry-points:
-      - 'src/liblsl-d6c43724/include/lsl_c.h'
+      - 'src/liblsl-846c4199/include/lsl_c.h'
     include-directives:
-      - 'src/liblsl-d6c43724/include/lsl/*.h'
-      - 'src/liblsl-d6c43724/include/lsl_c.h'
+      - 'src/liblsl-846c4199/include/lsl/*.h'
+      - 'src/liblsl-846c4199/include/lsl_c.h'
   compiler-opts:
-    - '-Isrc/liblsl-d6c43724/include'
+    - '-Isrc/liblsl-846c4199/include'
   structs:
     dependency-only: opaque
   functions:
