+++ packages/liblsl/pubspec.yaml
@@ -12,15 +12,15 @@ environment:
 
 dependencies:
   ffi: ^2.1.4
-  hooks: ^0.20.5
+  hooks: ^1.0.0
   native_toolchain_c: ^0.17.2
-  code_assets: ^0.19.10
+  code_assets: ^1.0.0
   logging: ^1.3.0
   meta: ^1.16.0
   fast_immutable_collections: ^11.1.0
 
 dev_dependencies:
-  ffigen: ^19.0.0
+  ffigen: ^20.0.0
   lints: ^6.0.0
   test: ^1.26.0
   dartdoc: ^9.0.0
