### 2025-11-24 94aac78
Added option for coordinator to participate in UserCoordinationMessages (via inlet).
