### 2026-02-03 eccff10
Updated JOSS paper to meet new section requirements.
