+++ packages/liblsl/paper/paper.md
@@ -38,31 +38,34 @@ affiliations:
   index: 3
   ror: 01aj84f44
 
-date: 1 October 2025
+date: 3 February 2026
 bibliography: paper.bib
 ---
 
 # Summary
 
-The `liblsl` Dart package is the first implementation of [Lab Streaming Layer (LSL)](https://labstreaminglayer.org/) in the [Dart](https://dart.dev/) and [Flutter](https://flutter.dev/) ecosystem, enabling researchers to deploy multi-modal LSL-enabled data acquisition applications across mobile (iOS, Android) and desktop (Linux, macOS, Windows) platforms from shared source code. Unlike [existing LSL implementations](https://labstreaminglayer.readthedocs.io/info/language_wrappers.html) that are platform-specific or restricted to desktop devices [@LiblslLanguageWrappers], this package leverages Dart/Flutter's cross-platform capabilities while maintaining the microsecond-level latency requirements of neurophysiological research through direct Foreign Function Interface (FFI) bindings to the native LSL C library [@stennerSccnLiblslV11622023].
-
+The `liblsl` Dart package is the first implementation of [Lab Streaming Layer (LSL)](https://labstreaminglayer.org/) in the [Dart](https://dart.dev/) and [Flutter](https://flutter.dev/) ecosystem, enabling researchers to deploy multi-modal LSL-enabled data acquisition applications across mobile (iOS, Android) and desktop (Linux, macOS, Windows) platforms from shared source code.
 
 
 # Statement of need
 
-Neuroscience and behavioral research increasingly make use of consumer hardware due to its lower cost and sufficient performance for running experiments [@roqueRealTimeMobileEEG2025], however, the integration of consumer devices into laboratory data acquisition pipelines often requires platform-specific development and adds complexity to the data alignment and collection process, both of which can hinder flexibility and reproducibility [@iwamaTwoCommonIssues2024]. The LSL system partially adresses some of the complexity via a software based unified method of time-synchronized data acquisition across heterogeneous hardware and software systems [@kotheLabStreamingLayer2025]. LSL handles clock synchronization, network communication, and data buffering, making it particularly valuable for applications requiring precise temporal alignment of multiple data sources. The `liblsl` Dart package further reduces the complexity by enabling researchers to deploy LSL-enabled applications across all major platforms from a single Dart/Flutter codebase. Potential use cases include general multimodal experiments [@wangScopingReviewUse2023; @dolmans2020data], mobile brain-computer interfaces using an EEG headset and a smartphone [@stopczynskiSmartphoneBrainScanner2014; @debenerUnobtrusiveAmbulatoryEEG2015; @blumEEGRecordingOnline2017a], hyperscanning studies [@zammPracticalGuideEEG2024b] that simultaneously collect neuroimaging data from multiple participants [@luftSocialSynchronizationBrain2022; @boggioSocialAffectiveNeuroscience2023; roqueRealTimeMobileEEG2025], and distributed experiments where multiple labs using a variety of devices collect methodologically consistent data in a standardized format for analysis [@demazureDistributedRemoteEEG2021; @schultzLinkingLabsInterconnecting2021]. 
[…]
