+++ packages/liblsl/paper/paper.md
@@ -38,7 +38,7 @@ affiliations:
   index: 3
   ror: 01aj84f44
 
-date: 3 February 2026
+date: 4 February 2026
 bibliography: paper.bib
 ---
 
@@ -47,13 +47,13 @@ bibliography: paper.bib
 The `liblsl` Dart package is the first implementation of [Lab Streaming Layer (LSL)](https://labstreaminglayer.org/) in the [Dart](https://dart.dev/) and [Flutter](https://flutter.dev/) ecosystem, enabling researchers to deploy multi-modal LSL-enabled data acquisition applications across mobile (iOS, Android) and desktop (Linux, macOS, Windows) platforms from shared source code.
 
 
-# Statement of need
+# Statement of Need
 
 Neuroscience and behavioral research increasingly make use of consumer hardware due to its lower cost and sufficient performance for running experiments [@roqueRealTimeMobileEEG2025], however, the integration of consumer devices into laboratory data acquisition pipelines often requires platform-specific development and adds complexity to the data alignment and collection process, both of which can hinder flexibility and reproducibility [@iwamaTwoCommonIssues2024]. The LSL system partially addresses some of the complexity via a software based unified method of time-synchronized data acquisition across heterogeneous hardware and software systems [@kotheLabStreamingLayer2025]. LSL handles clock synchronization, network communication, and data buffering, making it particularly valuable for applications requiring precise temporal alignment of multiple data sources. The `liblsl` Dart package further reduces the complexity by enabling researchers to deploy LSL-enabled applications across all major platforms from a single Dart/Flutter codebase. Potential use cases include general multimodal experiments [@wangScopingReviewUse2023; @dolmans2020data], mobile brain-computer interfaces using an EEG headset and a smartphone [@stopczynskiSmartphoneBrainScanner2014; @debenerUnobtrusiveAmbulatoryEEG2015; @blumEEGRecordingOnline2017a], hyperscanning studies [@zammPracticalGuideEEG2024b] that simultaneously collect neuroimaging data from multiple participants [@luftSocialSynchronizationBrain2022; @boggioSocialAffectiveNeuroscience2023; @roqueRealTimeMobileEEG2025], and distributed experiments where multiple labs using a variety of devices collect methodologically consistent data in a standardized format for analysis [@demazureDistributedRemoteEEG2021; @schultzLinkingLabsInterconnecting2021]. 
 
-# State of the field
+# State of the Field
 
- There are [existing LSL implementations](https://labstreaminglayer.readthedocs.io/info/language_wrappers.html) in other programming languages that are platform-specific or restricted to desktop devices [@LiblslLanguageWrappers], this package leverages Dart/Flutter's cross-platform capabilities while maintaining the microsecond-level latency requirements of neurophysiological research through direct Foreign Function Interface (FFI) bindings to the native LSL C library [@stennerSccnLiblslV11622023]. The need for a Dart implementation of `liblsl` arose during development of a group multimodal experiment paradigm that is intended to become available as a research tool, as such, making it available to as wide an audience as possible makes Dart a logical choice because distributable applications can readily be built for all major desktop and mobile operating systems.
+ There are [existing LSL implementations](https://labstreaminglayer.readthedocs.io/info/language_wrappers.html) in other programming languages that are platform-specific or restricted to desktop devices [@LiblslLanguageWrappers], this package leverages Dart/Flutter's cross-platform capabilities while maintaining the millisecond level latency requirements of neurophysiological research through direct Foreign Function Interface (FFI) bindings to the native LSL C library [@stennerSccnLiblslV11622023]. The need for a Dart LSL package is particularly extant within the behavioural science research community, where researchers often require multiplatform applications but lack the ability to develop applications for new platforms.
 
 # Software design
 
[…]
