+++ packages/liblsl_coordinator/CHANGELOG.md
@@ -1,3 +1,3 @@
-## 0.1.0
+## 0.1.1+0
 
 - Initial version.
