### 2026-02-09 15ff3ce
Update liblsl submodule, change build hook to match updated spec.
