+++ packages/liblsl/pubspec.yaml
@@ -11,19 +11,19 @@ environment:
   sdk: ^3.9.0
 
 dependencies:
-  ffi: ^2.1.4
-  hooks: ^1.0.0
-  native_toolchain_c: ^0.17.2
+  ffi: ^2.1.5
+  hooks: ^1.0.1
+  native_toolchain_c: ^0.17.4
   code_assets: ^1.0.0
   logging: ^1.3.0
-  meta: ^1.16.0
+  meta: ^1.18.0
   fast_immutable_collections: ^11.1.0
 
 dev_dependencies:
-  ffigen: ^20.0.0
-  lints: ^6.0.0
-  test: ^1.26.0
-  dartdoc: ^9.0.0
+  ffigen: ^20.1.1
+  lints: ^6.1.0
+  test: ^1.29.0
+  dartdoc: ^9.0.2
 
 ffigen:
   name: Liblsl
