+++ packages/liblsl_analysis/pubspec.yaml
@@ -39,11 +39,11 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
-  file_picker: ^10.1.9
+  file_picker: ^10.3.10
   csv: ^6.0.0
-  dartframe: ^0.8.5
-  fl_chart: ^1.0.0
-  shared_preferences: ^2.5.3
+  dartframe: ^0.8.10
+  fl_chart: ^1.1.1
+  shared_preferences: ^2.5.4
 
 dev_dependencies:
   flutter_test:
