### 2026-02-22 #9352 @sneakers-the-rat
Sure, i havent written Dart, but this looks broadly in my domain and i definitely understand the context/field, and skimming the code briefly i think i can read it enough to facilitate a review
