### 2026-02-23 #9352 @sneakers-the-rat
I'll take a more considered look tomorrow, but one initial question I have for the authors is how reviewers might test this/what they would need to evaluate the package(s) - if i'm not mistaken, typically this protocol is to facilitate streams across devices. I see some example code creating virtual streams here - https://github.com/NexusDynamic/liblsl.dart/blob/main/packages/liblsl/example/liblsl_example.dart - but since this is a language implementation of a cross-language protocol, it seems like reviewers might want to evaluate how it runs with e.g. devices that run on a protocol implementation in another language, etc. I see mentions in the docs about it running on a raspberry pi, android, ios, etc. are there e.g. test apps that they could use to evaluate cross-device streaming locally? Or could you set up some lsl server with some fake (or real) devices they could ping commands off of, stream from, etc.?

I'm just wondering what the phase of reviewing would look like where reviewers might want to try the tool out, run some examples, etc. and want to make sure we provide them with the resources/information they would need to do that.
