### 2026-02-23 #9352 @zeyus
> I'll take a more considered look tomorrow, but one initial question I have for the authors is how reviewers might test this/what they would need to evaluate the package(s) - if i'm not mistaken, typically this protocol is to facilitate streams across devices. I see some example code creating virtual streams here - https://github.com/NexusDynamic/liblsl.dart/blob/main/packages/liblsl/example/liblsl_example.dart - but since this is a language implementation of a cross-language protocol, it seems like reviewers might want to evaluate how it runs with e.g. devices that run on a protocol implementation in another language, etc. I see mentions in the docs about it running on a raspberry pi, android, ios, etc. are there e.g. test apps that they could use to evaluate cross-device streaming locally? Or could you set up some lsl server with some fake (or real) devices they could ping commands off of, stream from, etc.?
>
> I'm just wondering what the phase of reviewing would look like where reviewers might want to try the tool out, run some examples, etc. and want to make sure we provide them with the resources/information they would need to do that.

Great questions.

So, if someone wanted to test this with multiple devices that is definitely possible. In fact the [liblsl_example.dart](https://github.com/NexusDynamic/liblsl.dart/blob/main/packages/liblsl/example/liblsl_example.dart) will do the following:
- Create a network stream (visible to any LSL enabled device on the network) and send some data
- List up to 5 found streams on the network (including from other devices)
- In the multi stream example, it creates 2 outlets (outgoing streams) and sends data, these are also available to other network devices

The automated test suite also does similar things, with one important difference:
- they are configured to only listen and bind to the local loopback address (127.0.0.1) see [liblsl_test.dart](https://github.com/NexusDynamic/liblsl.dart/blob/main/packages/liblsl/test/liblsl_test.dart#L16-L17).
This means that streams using the automated test will not be visible to other devices, nor will streams from other devices be visible to the tests. Any of the code in there would work on the network if the API config was set to the defaults, or a different, less restrictive value).

If you were thinking along the lines of live viewing LSL stream content, that's definitely possible, but it would probably belong in a separate package (e.g. a cross-platform LSL live viewer). Otherwise if you were thinking along the lines of an app that just publishes and streams until you stop it (visible to other devices), I can definitely add an example app for that.

Actually, I might add in a producer and a consumer example that could be run on different devices.

*Edit: by the above, I mean, an example that just produces and publishes to a stream, and a consumer app that just lists all streams and lets you connect to one and show the data. That means either stream production, or consumption could be done with dart or any other language / device with LSL*

Because I am genuinely interested in making this package (and related suite) as accessible as possible, I'd be happy to add extra examples / apps to help with this. Including an emulated multi-device setup, but from experience, there are (operating system dependent) gotchas to do with multicast networking and packet forwarding within VMs, which would make this a good case for podman/docker-compose, except from a user perspective, I think that adds complexity and there would be nothing interactive to do with a bunch of services streaming to each other. Open to suggestions though!
