### 2026-02-23 #9352 @sneakers-the-rat
Since i dont write dart/flutter, ill need you to help me out a bit with knowing what a normal dart/flutter dev should be expected to know how to do :).

Sitting down to take a closer look and do editorial checks now, but on this point, the main thing im trying to clarify are methods by which reviewers can validate the functional claims made in the paper. This is a *tiny* bit more complicated than a "typical" package because it relates to a networking protocol, and the paper makes claims about being able to do e.g. multimodal research using smartphones (which we *may* expect reviewers to have access to) or other specialized devices (e.g. EEGs, which reviewers almost certainly won't have access to).

Basically we want to make sure that reviewers can read the paper, read the code, run the code, and confirm that it does what the paper says without them needing to e.g. write additional code or do any unduly burdensome setup of other devices. Trying to get this out of the way before inviting reviewers so they dont have to wait on hold.

So! Plenty of options for that, and again ill need you to let me know what is reasonable to expect reviewers to do and what would be possible for you to do.

E.g. in the past, a previous submission that was about some distributed network protocol where one of the nodes was particularly cumbersome or difficult to set up, the authors created a private endpoint for the reviewers to use during their review. So maybe you could set up some virtual "echo" device that reviewers could stream to and from?

Would it be reasonable to expect a dart/flutter dev to be able to build and deploy a demo app to their own phone? If so, could you provide the source of some minimal example app, and ideally ready-to-install binaries along with that? I dont think this needs to be particularly sophisticated or polished, since the submission is about a language adapter to a streaming protocol rather than an app - just enough for a reviewer to easily evaluate the ability of the submission to be used as described, to stream events between devices.

And yes, if it is typically necessary to use containers or VMs to run the software, then we would need to provide that & usage instructions to them.

If we get a basic idea of a setup, then we can work out the fine details with the reviewers after they're invited, not trying to unduly burden you as well, again we just want to a) have a clear picture of what the reviewers need so they know what they're getting into, and b) avoid lengthy back and forth spending reviewer time to figure out how to get a MWE running to evaluate the paper. So just let me know what you think would be the most reasonable path by which a reviewer might be able to confirm the claims of the paper, whether thats already in place or some additional scaffolding is needed :)
