### 2026-02-23 #9352 @sneakers-the-rat
other than that, quick spot checks looking at the package, it looks reviewable to me! though I imagine reviewers will probably immediately have questions about docs.

only other question I have right now is probably just me being unfamiliar with dart, but wondering about scope of review - is only the `liblsl` package the target of the review? (i.e. this directory: https://github.com/NexusDynamic/liblsl.dart/tree/main/packages/liblsl ) or are the other packages `liblsl_analysis`, `_coordinator`, `_test` and `_timing` also intended for review?
