### 2026-02-23 #9352 @zeyus
> though I imagine reviewers will probably immediately have questions about docs.

Fair, and to your previous comment, I will think about this and get back to you. But as an extremely imprecise summary until then: All the pieces should in place for testing (on any device) already, but I think it would be valuable to have a simple pre-compiled binary for various platforms (if people use iOS, I can publish a testflght app). In addition, to help reviewers, I will create a separate document and link to it in the readme which could act like a run-sheet which should be explicit enough for someone with little or no Dart experience should be able to follow (presuming they have familiarity with other programming languages or environments). I also will try to avoid Docker and other setups like that if possible because it is not relevant or necessary to use this package.

> only other question I have right now is probably just me being unfamiliar with dart, but wondering about scope of review - is only the liblsl package the target of the review?

@sneakers-the-rat This is something I was hoping yourself or another editor could shine light on. I have tried to find out how to fit this in to the JOSS process but it's still a little unclear:

- This paper is only related to the directory you linked to (the liblsl package)
- The liblsl package is the core, and the rest are built around it but are out of scope for this paper / review
- The liblsl.dart repository is a monorepo (a somewhat disliked, but fairly common pattern) for the following reasons:
  - Any other package there needs to be working with the core LSL functionality
  - This project is designed for researchers, and thus the other tools (timing tests, data analysis, etc) are intended as a suite to both set up and validate a lab / experimental setup
  - Despite this, it is possible and simple for developers to only install the core liblsl package and build software using it
  - Testing ensures the entire suite of tools remain functional and do not break with changes

While as I mentioned it's somewhat common (e.g. Flame, a Flutter-based game engine), I haven't seen any relevant JOSS submissions or documentation on how to handle this exactly, especially if in future I would like to submit the coordinator package (not relevant here, but would be core functionality for making experiments).

Anyway sorry for the wall of text, I would love to make this easy for everyone involved, so if a monorepo structure is not appropriate, I can consider breaking out the various packages into individual repositories.
