### 2026-02-24 #9352 @sneakers-the-rat
Will let you and the reviewers sort out what is to be expected with docs, I don't want to pre-set expectations, the joss checklist language is

>  Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

so since your package is a language adapter for an existing protocol, I don't know in advance to what degree reviewers will expect additional documentation beyond the protocol docs -- in part because i don't know what is self explanatory within dart and what would need documenting.

---

Monorepos are fine, main thing is we need to communicate clearly to reviewers what they are expected to review.

In my opinion the library by itself fits scope, I would see that as the thing being submitted, as that's also what the paper is about. I am not sure what the status/scope of the other packages is, but on cursory inspection I would sort of doubt that they would by themselves pass scope check. It is your submission, so it is up to you whether you want to make the other packages part of it - more is not always better, the submission is subject to the same brevity constraints, and reviewers can only be asked to review so much code; but I understand wanting to include all your hard work in the published work! To me they sort of read as example applications? Examples are themselves part of the review, so it might be good to label them as that. I *believe* resubmissions are allowed if the new work since the last submission itself passes scope check, but if that influences your decision I can check with EiC because I'm not exactly sure.
