### 2026-02-25 #9352 @zeyus
> but on cursory inspection I would sort of doubt that they would by themselves pass scope check.

I agree. I definitely do not want the other packages to be included as part of this submission. The timing test is functional but not structured or documented to a level that I would consider ready for public consumption. The core package should be sufficiently documented.

Thanks for the clarification.

Specifically for dart:
- Usually the example(s) serves as a reference
- The API documentation (Auto-generated and published on the package page) provides details about specific aspects of the code (provided the docblocks are set up and written correctly, which I have done)

But, I am happy to discuss with the reviewers ways to test the package that make it as frictionless as possible for them
