### 2026-02-25 6cf634e
bump liblsl_coordinator version.
