### 2026-03-02 8429857
[liblsl] add WIP review / testing guide
