### 2026-03-02 c87ee40
[liblsl] Increase test timeout on a flaky test. Add flaky tag to (now fixed) fullinfo test.
