+++ packages/liblsl/dart_test.yaml
@@ -1,3 +1,4 @@
 tags:
   performance:
     timeout: 10m
+  flaky:
