### 2026-03-02 e0813e3
[liblsl] Version bump, update liblsl submodule path. Add more explicit parameter type to getFullInfo isolate handler.
