### 2026-03-02 0a5d7a2
[liblsl_test] Added app icons (flutter_launcher_icons).
