+++ packages/liblsl_test/pubspec.yaml
@@ -38,7 +38,7 @@ dependencies:
     path: ../liblsl
   flutter:
     sdk: flutter
-  flutter_multicast_lock: ^1.1.0
+  flutter_multicast_lock: ^1.2.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -56,6 +56,7 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^6.0.0
+  flutter_launcher_icons: ^0.14.4
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
