### 2026-03-02 fa3e152
[liblsl] add WIP review / testing guide
