### 2026-03-03 2069bae
[liblsl] Updated testing guide
