+++ packages/liblsl/REVIEW_TESTING.md
@@ -3,11 +3,21 @@
 This is just a rough guide for testing options for this package. It will give some ideas based on which devices you have access to.
 
 <!-- vscode-markdown-toc -->
-1. [Initial setup](#Initialsetup)
-   * 1.1. [Getting the code](#Gettingthecode)
-   * 1.2. [Installing Dart and Flutter](#InstallingDartandFlutter)
-     * 1.2.1. [Option 1: FVM](#Option1:FVM)
-     * 1.2.2. [Option 2: Manual installation](#Option2:Manualinstallation)
+* 1\. [Initial setup](#Initialsetup)
+	* 1.1. [Getting the code](#Gettingthecode)
+	* 1.2. [Installing Dart and Flutter](#InstallingDartandFlutter)
+		* 1.2.1. [Option 1: FVM](#Option1:FVM)
+		* 1.2.2. [Option 2: Manual installation](#Option2:Manualinstallation)
+	* 1.3. [Preparing the environment](#Preparingtheenvironment)
+* 2\. [Testing process](#Testingprocess)
+	* 2.1. [First time setup](#Firsttimesetup)
+	* 2.2. [Automated tests](#Automatedtests)
+* 3\. [Integration testing](#Integrationtesting)
+	* 3.1. [liblsl_test example application for testing](#liblsl_testexampleapplicationfortesting)
+		* 3.1.1. [Running the precompiled binaries](#Runningtheprecompiledbinaries)
+	* 3.2. [Testing with another LSL enabled device or application](#TestingwithanotherLSLenableddeviceorapplication)
+		* 3.2.1. [Using a `Stream Viewer` application](#UsingaStreamViewerapplication)
+* 4\. [Anything else?](#Anythingelse)
 
 <!-- vscode-markdown-toc-config
 	numbering=true
@@ -44,20 +54,20 @@ The version of flutter is expected to be beta or dev/main[^1].
   - If you used the beta release, you can also switch to the dev channel by running `flutter channel main` and then `flutter upgrade`.
 
 
-### Preparing the environment
+###  1.3. <a name='Preparingtheenvironment'></a>Preparing the environment
 
 There is little beyond the normal Dart/Flutter setup required, but you should have clang installed and available in your PATH, as this is required for compiling the C code and creating the shared library. You can check if clang is available by running `clang --version` in your terminal.
 LLVM / Clang can be installed from here: https://github.com/llvm/llvm-project/releases/tag/llvmorg-18.1.8 but there is probably a version available through your package manager (e.g. `apt`, `pacman`, `brew`, `choco`, `winget`, etc.) that you can install.
 
 
-## Testing process
+##  2. <a name='Testingprocess'></a>Testing process
 
-### First time setup
+###  2.1. <a name='Firsttimesetup'></a>First time setup
 
 - Run `flutter pub get` in the repository root to get the dependencies.
 - Switch to the `packages/liblsl` directory, as the other packages are not relevant for testing the liblsl package.
 
-### Automated tests
+###  2.2. <a name='Automatedtests'></a>Automated tests
 
 There is a suite of automated tests that are included that test for core functionality. These explicitly only bind to the local / loopback interface, this means that any test streams will not be visible on your LAN (nor will it cause any traffic increase).
 
@@ -79,13 +89,42 @@ The automated tests consist of the following:
 
 - `test/liblsl_test.dart`: This tests all the core API functionality, such as creating streams, sending data, receiving data. It also tests the FFI bindings directly to ensure that they are working correctly.
 
-## Integration testing
+##  3. <a name='Integrationtesting'></a>Integration testing
 
[…]
