### 2026-03-03 00841cb
[liblsl] fix incompatible anchors in ghfm.
