### 2026-03-03 4950ac0
[liblsl] Replace TOC generator with github compatible version.
