+++ packages/liblsl/REVIEW_TESTING.md
@@ -1,43 +1,41 @@
 # Outline for testing the Dart liblsl package
 
 This is just a rough guide for testing options for this package. It will give some ideas based on which devices you have access to.
+<!-- TOC -->
 
-<!-- vscode-markdown-toc -->
-* 1\. [Initial setup](#Initialsetup)
-	* 1.1. [Getting the code](#Gettingthecode)
-	* 1.2. [Installing Dart and Flutter](#InstallingDartandFlutter)
-		* 1.2.1. [Option 1: FVM](#Option1FVM)
-		* 1.2.2. [Option 2: Manual installation](#Option2Manualinstallation)
-	* 1.3. [Preparing the environment](#Preparingtheenvironment)
-* 2\. [Testing process](#Testingprocess)
-	* 2.1. [First time setup](#Firsttimesetup)
-	* 2.2. [Automated tests](#Automatedtests)
-* 3\. [Integration testing](#Integrationtesting)
-	* 3.1. [liblsl_test example application for testing](#liblsl_testexampleapplicationfortesting)
-		* 3.1.1. [Running the precompiled binaries](#Runningtheprecompiledbinaries)
-	* 3.2. [Testing with another LSL enabled device or application](#TestingwithanotherLSLenableddeviceorapplication)
-		* 3.2.1. [Using a `Stream Viewer` application](#UsingaStreamViewerapplication)
-* 4\. [Anything else?](#Anythingelse)
+- [Outline for testing the Dart liblsl package](#outline-for-testing-the-dart-liblsl-package)
+    - [Initial setup](#initial-setup)
+        - [Getting the code](#getting-the-code)
+        - [Installing Dart and Flutter](#installing-dart-and-flutter)
+            - [Option 1: FVM](#option-1-fvm)
+            - [Option 2: Manual installation](#option-2-manual-installation)
+        - [Preparing the environment](#preparing-the-environment)
+    - [Testing process](#testing-process)
+        - [First time setup](#first-time-setup)
+        - [Automated tests](#automated-tests)
+    - [Integration testing](#integration-testing)
+        - [liblsl_test example application for testing](#liblsl_test-example-application-for-testing)
+            - [Running the precompiled binaries](#running-the-precompiled-binaries)
+        - [Testing with another LSL enabled device or application](#testing-with-another-lsl-enabled-device-or-application)
+            - [Using a Stream Viewer application](#using-a-stream-viewer-application)
+    - [Anything else?](#anything-else)
+- [Footnotes](#footnotes)
 
-<!-- vscode-markdown-toc-config
-	numbering=true
-	autoSave=true
-	/vscode-markdown-toc-config -->
-<!-- /vscode-markdown-toc -->
+<!-- /TOC -->
 
-##  1. <a name='Initialsetup'></a>Initial setup
+## Initial setup
 
 **Note: Flutter is not required for using the liblsl package, but if you want to run some of the example apps, then you will need to have Flutter installed.**
 
-###  1.1. <a name='Gettingthecode'></a>Getting the code
+### Getting the code
 
 - Clone the repository: `git clone --recurse-submodules https://github.com/NexusDynamic/liblsl.dart.git`
 
 *Note the `--recurse-submodules` flag, this is required because the C liblsl code is included as a git submodule.*
 
-###  1.2. <a name='InstallingDartandFlutter'></a>Installing Dart and Flutter
+### Installing Dart and Flutter
 
-####  1.2.1. <a name='Option1FVM'></a>Option 1: FVM
+#### Option 1: FVM
 
 FVM is probably the easiest way to go here both if you do not have an existing Dart/Flutter setup and also if you have other projects, this will help avoid version conflicts.
 
[…]
