### 2026-03-03 700e85e
[liblsl] Add screenshot placeholder.
