### 2026-03-03 3e86732
[liblsl] Update REVIEW_TESTING.md with application screenshots
