### 2026-03-03 #9352 @zeyus
@sneakers-the-rat sorry for the delay, I had to run some pilots and finish things off.

I have now updated the `liblsl_test` app (not part of the review, but a useful tool for the purposes of the review), to have a simple UI that allows all of the basic stream production / consumption functionality, and that will integrate with other instances of that app on connected devices, as well as any existing LSL enabled devices. For this app I have created application binaries for Linux, MacOS, Windows and Android (iOS is possible but requires significant time, and publishing to the app store). The following release contains the binaries for the `liblsl_test` app: https://github.com/NexusDynamic/liblsl.dart/releases/tag/liblsl_test_preview

I have also created a review / test guide which hopefully is of use for reviewers. That lives here: https://github.com/NexusDynamic/liblsl.dart/blob/main/packages/liblsl/REVIEW_TESTING.md
