### 2026-03-03 0ab48be
[liblsl] Clarification of Dart SDK version footnote
