### 2026-03-10 #9352 @sneakers-the-rat
Excellent!!!! Ok i'll let you communicate that and other details of what is needed to review the package with the reviewers once we find them. let's get this review going :)

I think we want at least one person experienced with dart, and I also typically think it's good to have someone who is in the expected audience of users. let's also see if we can find someone with prior expertise in streaming data or mobile dev (since, if i'm not mistaken, that's the main target of dart?)

Of the folks you had [listed above](https://github.com/openjournals/joss-reviews/issues/9352#issuecomment-3491243919), it seems like @monikasoni has relevant expertise, working across languages on IoT and performance sensitive streaming data. @monikasoni would you be willing to review this package?

Judging from the LSL docs and the paper, it seems like EEG is one of the main applications of LSL? I see @moan0s had listed LSL and EEG in their areas of expertise under a prior username, would you be willing to review this package? (also please let me know if your old username should be removed from the reviewer list)

If neither of these folks have Dart experience/are willing to review, I'll look for a dart person who can maybe do a half-review focusing on language idioms, even if they don't work on related application areas.
