### 2026-03-23 4f5e842
Use android_libcpp_shared package instead of using non-public native APIs.
