+++ packages/liblsl/pubspec.yaml
@@ -18,6 +18,7 @@ dependencies:
   logging: ^1.3.0
   meta: ^1.18.0
   fast_immutable_collections: ^11.1.0
+  android_libcpp_shared: ^0.1.0
 
 dev_dependencies:
   ffigen: ^20.1.1
