### 2026-04-13 #9352 @sneakers-the-rat
Sorry for the delay, my notifications exploded and i'm just catching up.

Looks like none of those reviewers are available, let's try another group.

From your list, I see @brpat07 who looks like a bit of a polyglot and may be able to review Dart and has contributed to the joss repo recently, @brpat07 would be willing to review this submission?

Other than that, @MikeAndroulakis is the only other person aside from you in our reviewer database that lists Dart or Flutter in their expertise, @MikeAndroulakis would you be willing to review?
