### 2026-04-14 #9352 @sneakers-the-rat
Totally reasonable - @zeyus would you be willing to commit to helping @brpat07 with understanding any language features that prove tricky? Sort of out of the norm, but we just don't have many (two, really) people in our reviewer pool who list Dart as being one of their languages. We want to try and minimize reviewer labor time if possible, since it's graciously volunteered, so rather than @brpat07 needing to spend time reviewing language docs and etc. it would be great if you would be willing to explain things as they come up even if they are language features rather than specific to the package.

fwiw I skimmed it and was able to get the gist, though of course reviewing is deeper reading.

Thank you @brpat07 for being willing to review! I'll add you in the next comment. That timeline is completely fine! You are volunteering your time, and we're grateful to have you.

I'll give a few more days for MikeAndroulakis to respond and then continue looking for reviewers!
