### 2026-04-14 #9352 @zeyus
@sneakers-the-rat  @brpat07 Thanks a lot for helping out here! I'm happy to help with any Dart-specific quirks. I have tried to make sure the library is well documented and understandable.

In general, Dart is quite nice to read, and most of the library is quite straightforward, with the exception being the use of [Isolates](https://dart.dev/language/isolates) to handle high-frequency concurrency. Dart uses the [Actor model](https://en.wikipedia.org/wiki/Actor_model) for threads. I'm happy to explain more if needed when you review, as it's the one place where it is slightly harder to trace the code during inter-thread communication.
