### 2026-04-17 #9352 @sneakers-the-rat
No problem, thanks for getting back to us @MikeAndroulakis !!! Safe travels.

So I think that strikes us out for people with listed Dart experience, and also with LSL experience, so i'm expanding search into those that list streaming and IoT expertise more generally.

@Vlado-V - I see you have listed streaming and wearables in the psychology space, would you be willing to review this package?

@Puneetha-Ramachandra has a large number of relevant expertise in streaming data and IoT frameworks, would you be willing to review this package?
