### 2026-04-17 #9352 @zeyus
@sneakers-the-rat just to let you know I've posted in the Dart Discord and on my Mastodon asking if people would like to help with JOSS reviews. My social media presence and influence is quite limited, but you never know, I signed up to be a reviewer from a Mastodon post by an editor.
