### 2026-04-17 #9352 @spydon
@zeyus I'd be happy to help out with reviewing!
Are there any specific guidelines for JOSS reviews?
