### 2026-04-17 #9352 @zeyus
@spydon Thanks so much, I really appreciate it!

Mostly it's following a checklist which will is generated for reviewers by @editorialbot (once you're assigned), the general guidelines are here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

@sneakers-the-rat what do you think can we get @spydon on board? :)
