### 2026-04-18 #9352 @sneakers-the-rat
Excellent :). Yes i'll add you both as reviewers. it's the end of the day for me, but i'll set a reminder for monday to come back and give reviewers a bit of an introduction if they need - the text in the issue that will be created will also be informative :). let's get this going
