### 2026-04-18 #10425 @zeyus
## Helpful documentation

- In addition to the normal code doc, readme, etc. I have made a guide to help with testing and understanding a bit about the package for reviewers. You can see this review and testing doc here: https://github.com/NexusDynamic/liblsl.dart/blob/main/packages/liblsl/REVIEW_TESTING.md
