+++ packages/liblsl/pubspec.yaml
@@ -1,6 +1,7 @@
 name: liblsl
-description: A Dart (and Flutter) native library for working with Lab Streaming Layer (LSL / liblsl).
-version: 0.11.0+0
+description: A Dart (and Flutter) native library for working with Lab Streaming
+  Layer (LSL / liblsl).
+version: 0.12.0+0
 homepage: https://zeyus.com/
 repository: https://github.com/NexusDynamic/liblsl.dart
 issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
@@ -16,7 +17,7 @@ dependencies:
   native_toolchain_c: ^0.17.4
   code_assets: ^1.0.0
   logging: ^1.3.0
-  meta: ^1.18.0
+  meta: ^1.17.0
   fast_immutable_collections: ^11.1.0
   android_libcpp_shared: ^0.1.0
 
@@ -31,18 +32,18 @@ ffigen:
   silence-enum-warning: true
   ffi-native:
   description: Bindings to `lsl_c.h`.
-  output: 'lib/native_liblsl.dart'
+  output: "lib/native_liblsl.dart"
   headers:
     entry-points:
-      - 'src/liblsl-9f0b6122/include/lsl_c.h'
+      - "src/liblsl-9f0b6122/include/lsl_c.h"
     include-directives:
-      - 'src/liblsl-9f0b6122/include/lsl/*.h'
-      - 'src/liblsl-9f0b6122/include/lsl_c.h'
+      - "src/liblsl-9f0b6122/include/lsl/*.h"
+      - "src/liblsl-9f0b6122/include/lsl_c.h"
   compiler-opts:
-    - '-Isrc/liblsl-9f0b6122/include'
+    - "-Isrc/liblsl-9f0b6122/include"
   structs:
     dependency-only: opaque
   functions:
     expose-typedefs:
       include:
-        - '.*'
+        - ".*"
