### 2026-04-22 #10425 @spydon
>  Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

@zeyus it would be good to have the standard `CONTRIBUTING.md` file in the repository.
https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors
