+++ packages/liblsl/README.md
@@ -6,7 +6,7 @@ This package provides a Dart wrapper for liblsl, using the dart native-assets bu
 
 Submitted [JOSS paper](./paper/paper.md): markdown version of the JOSS paper.
 
-[Take a look at the review/test guid](./REVIEW_TESTING.md) for testing options and instructions.
+[Take a look at the review/test guide](./REVIEW_TESTING.md) for testing options and instructions.
 
 ## Targets
 
@@ -233,11 +233,7 @@ dart test
 
 ## Contributing
 
-1. Fork the repository
-2. Create a feature branch (`git checkout -b feature/amazing-feature`)
-3. Commit your changes (`git commit -m 'Add amazing feature'`)
-4. Push to the branch (`git push origin feature/amazing-feature`)
-5. Open a Pull Request
+See the [CONTRIBUTING.md](./CONTRIBUTING.md) file for guidelines on how to contribute to this project.
 
 ### Development Guidelines
 
