### 2026-04-22 cdbc5aa
Add documents as copies rather than symlink to avoid confusion. These should not change too often.
