+++ packages/liblsl/SUPPORT.md
@@ -1,1 +1,14 @@
-../../SUPPORT.md
+# Getting support for liblsl.dart
+
+Liblsl.dart is a community-driven project, a lot of the core functionality is based around the upstream C/C++ [liblsl](https://github.com/sccn/liblsl) library.
+
+If you have any questions, issues or feature requests, feel free to open an [issue](https://github.com/NexusDynamic/liblsl.dart/issues) on GitHub.
+
+For more general support or discussion you may also use the following resources:
+
+- [GitHub Discussions](https://github.com/NexusDynamic/liblsl.dart/discussions)
+- [Matrix #liblsl.dart:neuro.wang](https://matrix.to/#/#liblsl.dart:neuro.wang)
+
+## Last Resort
+
+The above options are strongly preferred, but if for any reason these do not work for you, you may try one of the other channels listed on https://me.zys.im/, including emailing dev\[at\]zeyus\[dot\]com, but please only use this as a backup option because there is a lot of spam to that domain and your message might be missed.
