### 2026-04-22 #10425 @zeyus
Good point @spydon , I have added `CONTRIBUTING.md`, `CODE_OF_CONDUCT.md`, `SECURITY.md` and `SUPPORT.md`. These are all in the root of the monorepo, and exist as copies in the `liblsl` package.
