+++ packages/liblsl/pubspec.yaml
@@ -1,7 +1,7 @@
 name: liblsl
 description: A Dart (and Flutter) native library for working with Lab Streaming
   Layer (LSL / liblsl).
-version: 0.12.0+0
+version: 0.13.0+0
 homepage: https://zeyus.com/
 repository: https://github.com/NexusDynamic/liblsl.dart
 issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
@@ -14,12 +14,12 @@ environment:
 dependencies:
   ffi: ^2.1.5
   hooks: ^1.0.1
-  native_toolchain_c: ^0.17.4
+  native_toolchain_c: ^0.18.0
   code_assets: ^1.0.0
   logging: ^1.3.0
   meta: ^1.17.0
   fast_immutable_collections: ^11.1.0
-  android_libcpp_shared: ^0.1.0
+  android_libcpp_shared: ^0.1.1
 
 dev_dependencies:
   ffigen: ^20.1.1
