+++ packages/liblsl_timing/pubspec.yaml
@@ -41,24 +41,23 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.8
+  cupertino_icons: ^1.0.9
   path_provider: ^2.1.5
-  csv: ^6.0.0
+  csv: ^8.0.0
   path: ^1.9.1
-  shared_preferences: ^2.5.3
-  file_picker: ^10.1.9
-  synchronized: ^3.3.1
+  shared_preferences: ^2.5.5
+  file_picker: ^11.0.2
+  synchronized: ^3.4.0+1
   easy_localization: ^3.0.8
   flutter_fullscreen: ^1.2.0
-  wakelock_plus: ^1.3.2
+  wakelock_plus: ^1.6.0
   collection: ^1.19.1
-  share_plus: ^12.0.0
-  network_info_plus: ^7.0.0
-  permission_handler: ^12.0.0+1
-  flutter_refresh_rate_control: ^0.0.3+1
-# This was needed for easy_localization < 3.0.8, but now it is not.
-# dependency_overrides:
-#   intl: ^0.20.2
+  share_plus: ^13.1.0
+  network_info_plus: ^8.1.0
+  permission_handler: ^12.0.1
+  flutter_refresh_rate_control: ^0.0.6+0
+dependency_overrides:
+  win32: ^6.0.1  # Override for file_picker
 dev_dependencies:
   flutter_test:
     sdk: flutter
