### 2026-04-22 ba2d3e4
Added matrix chat link to readme.
