### 2026-05-13 82df3db
Added dockerfile and instructions for running paper analysis script.
