+++ packages/liblsl/paper/analysis/latency_analysis.R
@@ -8,20 +8,56 @@ library(conflicted)
 library(ggprism)
 library(showtext)
 library(knitr)
+library(progress)
+
+
+spawn_progressbar <- function(x, .name = .pb, .times = 1) {
+    .name <- substitute(.name)
+    n <- nrow(x) * .times
+    eval(
+        substitute(
+            .name <<- progress_bar$new(
+                format = "  processing [:bar] :percent eta: :eta",
+                total = n,
+                clear = FALSE,
+                width = 60
+            )
+        )
+    )
+    .pb$tick(0)
+    x
+}
 
 ####### SETUP #######
-# install the New Computer Modern Sans font from:
-# https://download.gnu.org.ua/release/newcm/
+# Font files are included in the fonts/ subdirectory (New Computer Modern Sans,
+# licensed under the GUST Font License — see fonts/LICENSE.txt).
+# If the files are not found the script falls back to the system "sans" font.
 jossFont <- "NewComputerModernSans10"
-jossFontFileName <- "NewCMSans10-Book.otf"
-jossFontFileNameBold <- "NewCMSans10-Bold.otf"
 
-font_add(
-    family = jossFont,
-    regular = jossFontFileName,
-    bold = jossFontFileNameBold,
-)
-showtext_auto() 
+.findFontFile <- function(filename) {
+    candidates <- c(
+        file.path("fonts", filename),   # repo layout: fonts/ next to the script
+        file.path("/fonts", filename),  # Docker image path
+        filename                        # working-directory fallback
+    )
+    found <- Filter(file.exists, candidates)
+    if (length(found)) found[[1]] else NULL
+}
+
+.font_regular <- .findFontFile("NewCMSans10-Book.otf")
+.font_bold    <- .findFontFile("NewCMSans10-Bold.otf")
+
+if (!is.null(.font_regular) && !is.null(.font_bold)) {
+    # add the font to the font search path
+    font_paths(c(dirname(.font_regular)))
+    font_add(family = jossFont, regular = .font_regular, bold = .font_bold)
+    showtext_auto()
+} else {
+    jossFont <- "sans"
+    message("NewComputerModernSans10 not found — falling back to 'sans'. ",
+            "Ensure fonts/NewCMSans10-Book.otf and fonts/NewCMSans10-Bold.otf ",
+            "are present for an exact reproduction of the published figure.")
+}
 
 ## Geom split violin from: https://stackoverflow.com/a/47652563/570122
 GeomSplitViolin <- ggproto("GeomSplitViolin", GeomViolin,
@@ -68,17 +104,25 @@ create_quantile_segment_frame <- function(data, draw_quantiles, split = FALSE, g
   }
 }
 
+parse_json_metadata <- function(x, .pb) {
+    json_data <- fromJSON(x)
+    .pb$tick()
+    # Convert NULL values to NA and ensure all elements are vectors
+    json_data[sapply(json_data, is.null)] <- NA
+    as_tibble(json_data)
+}
+
 
 ##### LOAD DATA ######
 
-geom_split_violin <- function(mapping = NULL, data = NULL, stat = "ydensity", position = "identity", ..., 
-                              draw_quantiles = NULL, trim = TRUE, scale = "area", na.rm = FALSE, 
+geom_split_violin <- function(mapping = NULL, data = NULL, stat = "ydensity", position = "identity", ...,
+                              draw_quantiles = NULL, trim = TRUE, scale = "area", na.rm = FALSE,
                               show.legend = NA, inherit.aes = TRUE) {
[…]
