+++ packages/liblsl/paper/paper.bib
@@ -74,7 +74,8 @@
   year = {2020},
   journal = {Measuring Behavior 2020-21 Volume},
   volume = {1},
-  pages = {26--32}
+  pages = {26--32},
+  doi = {10.6084/m9.figshare.13013717}
 }
 
 @article{iwamaTwoCommonIssues2024,
