### 2026-05-14 e2016de
Fix #21 - URL render issue in bibliography.
