+++ packages/liblsl/paper/paper.bib
@@ -12,8 +12,7 @@
   urldate = {2025-10-10},
   abstract = {Objective. Our aim was the development and validation of a modular signal processing and classification application enabling online electroencephalography (EEG) signal processing on off-the-shelf mobile Android devices. The software application SCALA (Signal ProCessing and CLassification on Android) supports a standardized communication interface to exchange information with external software and hardware. Approach. In order to implement a closed-loop brain-computer interface (BCI) on the smartphone, we used a multiapp framework, which integrates applications for stimulus presentation, data acquisition, data processing, classification, and delivery of feedback to the user. Main Results. We have implemented the open source signal processing application SCALA. We present timing test results supporting sufficient temporal precision of audio events. We also validate SCALA with a well-established auditory selective attention paradigm and report above chance level classification results for all participants. Regarding the 24-channel EEG signal quality, evaluation results confirm typical sound onset auditory evoked potentials as well as cognitive event-related potentials that differentiate between correct and incorrect task performance feedback. Significance. We present a fully smartphone-operated, modular closed-loop BCI system that can be combined with different EEG amplifiers and can easily implement other paradigms.},
   copyright = {Copyright {\copyright} 2017 Sarah Blum et al.},
-  langid = {english},
-  file = {/Users/au662726/Zotero/storage/TBJ9YUTV/Blum et al. - 2017 - EEG Recording and Online Signal Processing on Android A Multiapp Framework for Brain-Computer Inter.pdf;/Users/au662726/Zotero/storage/APKUUR6R/3072870.html}
+  langid = {english}
 }
 
 @book{boggioSocialAffectiveNeuroscience2023,
@@ -27,8 +26,7 @@
   urldate = {2025-10-10},
   copyright = {https://creativecommons.org/licenses/by/4.0},
   isbn = {978-3-031-08650-2 978-3-031-08651-9},
-  langid = {english},
-  file = {/Users/au662726/Zotero/storage/LAGSP4VC/Boggio et al. - 2023 - Social and Affective Neuroscience of Everyday Human Interaction From Theory to Methodology.pdf}
+  langid = {english}
 }
 
 @article{debenerUnobtrusiveAmbulatoryEEG2015,
@@ -47,8 +45,7 @@
   abstract = {This study presents first evidence that reliable EEG data can be recorded with a new cEEGrid electrode array, which consists of ten electrodes printed on flexible sheet and arranged in a c-shape to fit around the ear. Ten participants wore two cEEGrid systems for at least seven hours. Using a smartphone for stimulus delivery and signal acquisition, resting EEG and auditory oddball data were collected in the morning and in the afternoon six to seven hours apart. Analysis of resting EEG data confirmed well-known spectral differences between eyes open and eyes closed conditions. The ERP results confirmed the predicted condition effects with significantly larger P300 amplitudes for target compared to standard tones and a high test-retest reliability of the P300 amplitude (r\,{$>$}\,=\,.74). Moreover, a linear classifier trained on data from the morning session revealed similar performance in classification accuracy for the morning and the afternoon sessions (both\,{$>$}\,70\%). These findings demonstrate the feasibility of concealed and comfortable brain activity acquisition over many hours.},
   copyright = {2015 The Author(s)},
   langid = {english},
-  keywords = {Neuroscience,Psychology},
-  file = {/Users/au662726/Zotero/storage/CY72XRUA/Debener et al. - 2015 - Unobtrusive ambulatory EEG using a smartphone and flexible printed electrodes around the ear.pdf}
+  keywords = {Neuroscience,Psychology}
 }
 
 @inproceedings{demazureDistributedRemoteEEG2021,
@@ -89,8 +86,7 @@
   issn = {01680102},
   doi = {10.1016/j.neures.2023.12.003},
   urldate = {2025-09-19},
-  langid = {english},
[…]
