### 2026-05-22 f912a49
Update code metadata, bump version and add zenodo-release workflow
