+++ packages/liblsl/pubspec.yaml
@@ -1,7 +1,7 @@
 name: liblsl
 description: A Dart (and Flutter) native library for working with Lab Streaming
   Layer (LSL / liblsl).
-version: 0.13.0+0
+version: 0.13.1+0
 homepage: https://zeyus.com/
 repository: https://github.com/NexusDynamic/liblsl.dart
 issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
