### 2026-05-22 1a51ad1
Update code metadata, bump version and add zenodo-release workflow
