+++ .github/workflows/zenodo-release.yml
@@ -1,0 +1,182 @@
+name: Zenodo Release (liblsl)
+
+on:
+  workflow_dispatch:
+    inputs:
+      version:
+        description: "Release version: match packages/liblsl/pubspec.yaml"
+        required: true
+        type: string
+
+jobs:
+  release:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+
+    steps:
+      - name: Checkout with submodules
+        uses: actions/checkout@v4
+        with:
+          submodules: recursive
+          fetch-depth: 0
+
+      - name: Set version outputs
+        id: version
+        run: |
+          VERSION="${{ inputs.version }}"
+          TAG="liblsl-v$VERSION"
+          echo "tag=$TAG" >> "$GITHUB_OUTPUT"
+          echo "version=$VERSION" >> "$GITHUB_OUTPUT"
+          echo "archive=liblsl-$VERSION.zip" >> "$GITHUB_OUTPUT"
+
+      - name: Verify pubspec version matches
+        run: |
+          PUBSPEC_VERSION=$(grep '^version:' packages/liblsl/pubspec.yaml | sed 's/version: *//' | sed 's/+.*//' | tr -d ' ')
+          if [ "$PUBSPEC_VERSION" != "${{ inputs.version }}" ]; then
+            echo "ERROR: pubspec.yaml version ($PUBSPEC_VERSION) does not match input (${{ inputs.version }})"
+            exit 1
+          fi
+
+      - name: Sync metadata file versions
+        env:
+          VERSION: ${{ steps.version.outputs.version }}
+        run: |
+          TODAY=$(date -u +%Y-%m-%d)
+          export TODAY
+
+          # CITATION.cff
+          sed -i "s/^version: .*/version: \"$VERSION\"/" CITATION.cff
+          sed -i "s/^date-released: .*/date-released: \"$TODAY\"/" CITATION.cff
+
+          # codemeta.json and .zenodo.json
+          python3 - <<'PYEOF'
+          import json, os
+
+          version = os.environ["VERSION"]
+          today = os.environ["TODAY"]
+
+          for path, keys in [
+              ("codemeta.json", {"version": version, "dateModified": today}),
+              (".zenodo.json",  {"version": version}),
+          ]:
+              with open(path) as f:
+                  data = json.load(f)
+              data.update(keys)
+              with open(path, "w") as f:
+                  json.dump(data, f, indent=2)
+                  f.write("\n")
+              print(f"Updated {path}")
+          PYEOF
+
+      - name: Commit code meta, tag
+        run: |
+          git config user.name "github-actions[bot]"
+          git config user.email "github-actions[bot]@users.noreply.github.com"
+          git add CITATION.cff codemeta.json .zenodo.json
+          if git diff --cached --quiet; then
+            echo "No metadata changes — creating tag on current HEAD."
+          else
+            git commit -m "chore: release liblsl ${{ steps.version.outputs.version }}"
+            git push origin HEAD:main
+          fi
+          git tag "${{ steps.version.outputs.tag }}"
+          git push origin "${{ steps.version.outputs.tag }}"
+
+      - name: Build release archive
+        run: |
+          VERSION="${{ steps.version.outputs.version }}"
+          ARCHIVE="${{ steps.version.outputs.archive }}"
+          STAGING="liblsl-${VERSION}"
+
+          mkdir -p "${STAGING}"
+
+          # Package source (includes paper/, src/, lib/, test/, etc.)
[…]
