+++ CITATION.cff
@@ -2,7 +2,7 @@ cff-version: 1.2.0
 message: "If you use this software, please cite it as below."
 type: software
 title: "liblsl.dart: A Dart native API for Lab Streaming Layer (LSL)"
-version: "0.13.1"
+version: "0.13.2"
 date-released: "2026-05-22"
 license: MIT
 repository-code: "https://github.com/NexusDynamic/liblsl.dart"
