### 2026-05-22 48601fe
Add zenodo doi to readme, add some basic setup instructions for the monorepo.
