### 2026-06-01 bc15bf3
Added coordinator lsl stream helper property "running"
