### 2026-08-03 eb9cdcf
Coordinator bugfixes
