+++ packages/liblsl_coordinator/pubspec.yaml
@@ -1,6 +1,6 @@
 name: liblsl_coordinator
 description: A performance-focused Dart (and Flutter) LSL-based device coordination library.
-version: 0.1.4+2
+version: 0.2.0+0
 homepage: https://zeyus.com/
 repository: https://github.com/NexusDynamic/liblsl.dart
 issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
@@ -11,11 +11,11 @@ publish_to: 'none'
 resolution: workspace
 
 environment:
-  sdk: ^3.11.0
+  sdk: ^3.12.2
 
 # Add regular dependencies here.
 dependencies:
-  meta: ^1.16.0
+  meta:
   # liblsl: ^0.9.0
   liblsl:
     path: ../liblsl
@@ -30,4 +30,4 @@ dev_dependencies:
   integration_test:
     sdk: flutter
   lints: ^6.0.0
-  test: ^1.24.0
+  test: ^1.31.2
