### 2026-08-03 9cbb7af
Performance improvements, upstream (beta) API parity
