+++ packages/liblsl_timing/pubspec.yaml
@@ -54,7 +54,7 @@ dependencies:
   collection: ^1.19.1
   share_plus: ^13.1.0
   network_info_plus: ^8.1.0
-  permission_handler: ^12.0.1
+  permission_handler: ^13.0.0
   flutter_refresh_rate_control: ^0.0.6+0
 dev_dependencies:
   flutter_test:
