+++ packages/liblsl_test/pubspec.yaml
@@ -25,7 +25,7 @@ resolution: workspace
 version: 1.1.1+1
 
 environment:
-  sdk: ^3.9.0
+  sdk: ^3.12.2
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -39,6 +39,7 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_multicast_lock: ^1.2.0
+  meta:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -57,6 +58,8 @@ dev_dependencies:
   # rules and activating additional ones.
   flutter_lints: ^6.0.0
   flutter_launcher_icons: ^0.14.4
+  test: ^1.31.2
+
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec

[…]
