### 2026-08-09 729c34b
WIP tests for coordinator
