### 2026-08-10 be06418
Seperated peer_coordinator and liblsl_coordinator
