+++ packages/peer_coordinator/pubspec.yaml
@@ -1,0 +1,39 @@
+name: peer_coordinator
+description: >-
+  Transport-neutral peer coordination: election, membership, heartbeat and
+  synchronised data streams over a pluggable backend. Pure Dart, web-safe.
+version: 0.1.0
+homepage: https://zeyus.com/
+repository: https://github.com/NexusDynamic/liblsl.dart
+issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
+# Not published yet; versioned alongside liblsl_coordinator for now.
+publish_to: 'none'
+
+resolution: workspace
+
+environment:
+  sdk: ^3.12.2
+
+dependencies:
+  meta:
+  synchronized: ^3.4.0+1
+  logging: ^1.3.0
+  uuid: ^4.5.3
+  collection: ^1.19.1
+  fast_immutable_collections: ^11.2.0
+  # Dart-team maintained; one API on the VM and in the browser. Only reached
+  # by `package:peer_coordinator/websocket.dart`, so an app that never imports
+  # that library never compiles it in.
+  web_socket_channel: ^3.0.3
+  # A regular dependency, not a dev one, because `lib/testing.dart` ships the
+  # cross-transport conformance suite for other packages to run against their
+  # own backend. Nothing in the runtime libraries imports it, so it is
+  # tree-shaken out of any real build.
+  test: ^1.31.2
+
+executables:
+  # dart run peer_coordinator:hub --port 8080
+  hub:
+
+dev_dependencies:
+  lints: ^6.0.0
