### 2026-08-10 fad8500
Added identities to inlet, outlet and streaminfo objects
