+++ packages/liblsl/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.14.1+0
+
+- Added identities to `LSLInlet`, `LSLOutlet` and `LSLStreamInfo` (`hashCode` and `==`).
+
 # 0.14.0+0
 
 ## New features
