### 2026-08-10 8ef127c
Added more tests and resolved some bugs during decoupling
