+++ packages/liblsl_coordinator/CHANGELOG.md
@@ -1,6 +1,143 @@
-## Unreleased
+## 0.4.0+0
+
+The coordination layer is now transport-neutral and lives in a new pure-Dart
+package, `peer_coordinator`. `liblsl_coordinator` is the Lab Streaming Layer
+transport for it, and re-exports the core so existing imports keep working.
+
+Two other transports ship in `peer_coordinator`: an in-memory one for testing
+and a WebSocket one (with a relay hub) that also runs in the browser.
+
+### Breaking changes
+
+- **The core moved to `package:peer_coordinator`.** Every library previously
+  under `package:liblsl_coordinator/` — `framework.dart`, `config.dart`,
+  `coordination.dart`, `data.dart`, `discovery.dart`, `interfaces.dart`,
+  `network.dart`, `logging.dart` — is now a one-line re-export of its
+  `peer_coordinator` counterpart. Existing imports continue to resolve
+  unchanged; prefer importing from `peer_coordinator` in new code.
+  `liblsl_coordinator` now depends on it.
+- **`LSLCoordinationSession` is a thin subclass of `PeerSession`.** The session
+  flow, election, heartbeat, membership and stream lifecycle are all in
+  `PeerSession` now. Return types are narrowed covariantly, so
+  `createDataStream`/`getDataStream` still hand back `LSLDataStream` and
+  `transport` still returns `LSLTransport`. No call-site change expected.
+- **`ITransportConfig` requires `createTransport()`.** Transport selection is a
+  method on the config rather than a registry, so an application never
+  compiles in a transport it does not name. Only affects custom transports.
+- **`ITransport` gained `streamFactory` and `createDiscovery`, and lost
+  `createStream`.** `createStream` had no callers. `ITransport` now also
+  implements `IResourceManager`, which it did in practice already.
+- **`NetworkStream` gained the lifecycle members that previously existed only
+  on `LSLStreamMixin`**: `started`, `start`, `stop`, `createOutlet`,
+  `recreateOutlet`, `addInlet`, `createInletsForNodes`, `updateNode`, plus
+  default `pauseStream`/`resumeStream`/`flushStreams`/`destroyStream`. Callers
+  no longer need a concrete LSL type to drive a stream.
+- **`resume()` was split.** `IPausable.resume()` takes no arguments; the
+  parameterised form is now `resumeWith({flushBeforeResume})`. The old
+  widened `resume({flushBeforeResume})` could never be called through the base
+  type.
+- **`createResolvedInletsForStream` renamed to `createInletsForNodes`** and
+  **`addInlet` now takes a `PeerHandle`**, not an `LSLStreamInfo`.
+- **Discovery is typed.** `LslDiscovery` implements `IDiscovery`;
+  `startDiscovery(predicate:)`/`stopDiscovery()` are now `start(query:)`/
+  `stop()`, taking a `DiscoveryQuery` instead of an XPath string. The LSL
+  transport compiles queries to XPath via `LslPredicateCompiler`, which emits
+  byte-identical predicates (golden-tested). `LSLStreamInfoHelper`'s string
+  builders remain for now.
+- **Discovery events renamed.** `StreamDiscoveredEvent` is now
+  `PeersDiscoveredEvent`, carrying `List<PeerHandle>` rather than
+  `List<StreamInfoResource>`.
+- **`Log.sendPort` is now `Log.forwardTo(callback)`** and
+  `Log.logIsolateMessage` is `Log.replayRecord` (old names deprecated). Logging
+  no longer imports `dart:isolate` or `dart:io`, so the core compiles for web.
+  The default logger name changed from `LSLCoordinator` to `PeerCoordinator`.
+- **Removed as unused:** `TransportStreamConfig` /
[…]
