### 2026-08-10 7f9496a
Added test+peer_coordinator to pubspec
