### 2026-08-15 2d15625
added coordination timing info
