### 2026-08-18 7abe300
TO TEST: ws implementation, auto close session on coordinator leave
