### 2026-08-18 99a3aa4
webrtc_coordinator skeleton - add removeInlet to peer_coordinator structure
