+++ packages/webrtc_coordinator_flutter/pubspec.yaml
@@ -1,0 +1,31 @@
+name: webrtc_coordinator_flutter
+description: >-
+  The flutter_webrtc binding for webrtc_coordinator. Supplies an RtcPeerAdapter
+  backed by real peer connections; all transport logic lives in the pure-Dart
+  webrtc_coordinator package.
+version: 0.1.0
+homepage: https://zeyus.com/
+repository: https://github.com/NexusDynamic/liblsl.dart
+issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
+publish_to: 'none'
+
+resolution: workspace
+
+environment:
+  sdk: ^3.12.2
+  flutter: ^3.27.1
+
+dependencies:
+  flutter:
+    sdk: flutter
+  flutter_webrtc: ^1.6.0
+  logging: ^1.3.0
+  webrtc_coordinator:
+    path: ../webrtc_coordinator
+  peer_coordinator:
+    path: ../peer_coordinator
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  lints: ^6.0.0
