### 2026-08-18 96065a5
[liblsl] Update test for leaks to use property resolver.
