+++ packages/liblsl/pubspec.yaml
@@ -12,10 +12,10 @@ environment:
   sdk: ^3.11.0
 
 dependencies:
-  ffi: ^2.1.5
-  hooks: ^2.1.0
-  native_toolchain_c: ^0.19.3
-  code_assets: ^1.0.0
+  ffi: ^2.2.0
+  hooks: ^2.2.0
+  native_toolchain_c: ^0.19.4
+  code_assets: ^2.0.0
   logging: ^1.3.0
   meta:
   fast_immutable_collections: ^11.1.0
@@ -26,6 +26,9 @@ dev_dependencies:
   lints: ^6.1.0
   test: ^1.31.0
   dartdoc: ^9.0.8
+
+dependency_overrides:
+  code_assets: ^1.0.0 # temporarily override to avoid issues with equality, breaking change with code_assets: ^2.0.0
 
 ffigen:
   name: Liblsl
