+++ packages/liblsl/pubspec.yaml
@@ -15,11 +15,11 @@ dependencies:
   ffi: ^2.2.0
   hooks: ^2.2.0
   native_toolchain_c: ^0.19.4
-  code_assets: ^2.0.0
+  code_assets: ^1.2.1
   logging: ^1.3.0
   meta:
   fast_immutable_collections: ^11.1.0
-  android_libcpp_shared: ^0.2.0
+  android_libcpp_shared: ^0.3.0
 
 dev_dependencies:
   ffigen: ^21.0.0
