+++ packages/liblsl/pubspec_overrides.yaml
@@ -1,1 +1,0 @@
-resolution:
