+++ packages/peer_coordinator/pubspec.yaml
@@ -15,20 +15,16 @@ environment:
   sdk: ^3.12.2
 
 dependencies:
-  meta:
+  meta: any
   synchronized: ^3.4.0+1
   logging: ^1.3.0
   uuid: ^4.5.3
   collection: ^1.19.1
   fast_immutable_collections: ^11.2.0
-  # Dart-team maintained; one API on the VM and in the browser. Only reached
-  # by `package:peer_coordinator/websocket.dart`, so an app that never imports
+  # An app that never imports peer_coordinator/websocket.dart
   # that library never compiles it in.
   web_socket_channel: ^3.0.3
-  # A regular dependency, not a dev one, because `lib/testing.dart` ships the
-  # cross-transport conformance suite for other packages to run against their
-  # own backend. Nothing in the runtime libraries imports it, so it is
-  # tree-shaken out of any real build.
+  # tree-shaken out of a real build.
   test: ^1.31.2
 
 executables:
