+++ packages/webrtc_coordinator/README.md
@@ -1,0 +1,161 @@
+# webrtc_coordinator
+
+A genuinely peer-to-peer transport for
+[`peer_coordinator`](../peer_coordinator): coordination and data over WebRTC
+data channels, with the hub demoted to discovery and connection setup.
+
+Pure Dart. It ships no WebRTC implementation — you supply an `RtcPeerAdapter`.
+[`webrtc_coordinator_flutter`](../webrtc_coordinator_flutter) provides one
+backed by `flutter_webrtc`; `package:webrtc_coordinator/testing.dart` provides a
+fake for headless tests.
+
+## Why
+
+The WebSocket transport relays every byte. The hub is on the data path by
+design, so every message and every sample costs two network hops. Here it costs
+one: peers exchange an offer and an answer through the hub, and everything
+after that goes directly between them.
+
+The hop count is the obvious win. The less obvious one is that a data channel
+can be **unreliable and unordered**, which a relay cannot offer at all — so a
+latency-critical sampling stream can stop paying for retransmissions it does
+not want.
+
+No core changes were needed for any of this. The abstraction was built for it:
+LSL is already peer-to-peer behind the same interfaces.
+
+## Using it
+
+```dart
+import 'package:peer_coordinator/peer_coordinator.dart';
+import 'package:webrtc_coordinator/transports/webrtc.dart';
+import 'package:webrtc_coordinator_flutter/webrtc_coordinator_flutter.dart';
+
+final config = CoordinationConfig(
+  name: 'my_session',
+  transportConfig: RtcTransportConfig(
+    hubUri: Uri.parse('ws://hub.local:8080'),
+    adapterFactory: flutterWebrtcAdapterFactory,
+    // Data streams only; the coordination stream is always reliable and
+    // ordered, because a lost election message splits the session.
+    dataOrdered: false,
+    dataMaxRetransmits: 1,
+  ),
+);
+```
+
+Run the same hub as the WebSocket transport:
+
+```sh
+dart run peer_coordinator:hub --host 0.0.0.0 --port 8080
+```
+
+## What still goes through the hub
+
+Discovery and signalling, and nothing else.
+
+| | Hub | Direct |
+|---|---|---|
+| Endpoint registration, peer queries, election | ✅ | |
+| WebRTC offers, answers, ICE candidates | ✅ | |
+| Coordination messages | | ✅ |
+| Data samples | | ✅ |
+
+`test/transports/rtc_transport_test.dart` pins this: after a full data exchange
+the hub's routing table is empty for every stream, so it could not have relayed
+a byte even if one had arrived.
+
+## How it is put together
+
+* **One `RTCPeerConnection` per peer pair**, keyed on node uId — the only
+  identifier stable across an election role change — and shared by every stream
+  between those two peers.
+* **One data channel per stream**, opened `negotiated: true` with an id derived
+  from the stream name (`rtcChannelIdFor`), so neither end has to exchange
+  channel metadata. Derivation collisions are asserted on at open time rather
+  than trusted.
+* **Glare** is resolved by rule, not negotiation: the lexicographically lower
+  node uId makes the offer.
+* **Routing is local.** The hub's `RelayRouting` becomes
+  `RtcMesh.subscribersFor`, filled by the `open` signal a subscriber sends. A
+  producer sends on exactly the channels whose far end asked for them.
+* **Samples reuse `WsSampleFrame` verbatim**, so the encoders and
+  `decodeChannels` stay shared and covered by the WebSocket transport's tests.
+  Its slot fields are written as zero and never read — a data channel identifies
+  its sender by being that channel.
+
+### The adapter seam
+
+`flutter_webrtc` cannot run in a headless `dart test` VM, so every call into it
+goes through `lib/src/rtc/rtc_adapter.dart`, with `FakeRtcPeerAdapter` on the
[…]
