+++ packages/peer_coordinator/example/pubspec.yaml
@@ -17,6 +17,11 @@ environment:
 dependencies:
   peer_coordinator:
     path: ../
+  # The WebRTC transport, so the same chat can run peer-to-peer with the hub
+  # demoted to discovery and signalling. Brings flutter_webrtc with it, which
+  # is why the pure-Dart half lives in a separate package.
+  webrtc_coordinator_flutter:
+    path: ../../webrtc_coordinator_flutter
   flutter:
     sdk: flutter
   logging: ^1.3.0
