### 2026-08-18 8f2a4dd
Handle no-op coordinator messages, add STUN server to example
