+++ packages/peer_coordinator/pubspec.yaml
@@ -16,6 +16,8 @@ environment:
 
 dependencies:
   meta: any
+  # HMAC-SHA256 for the hub auth handshake. Pure Dart, so it stays web-safe.
+  crypto: ^3.0.6
   synchronized: ^3.4.0+1
   logging: ^1.3.0
   uuid: ^4.5.3
