### 2026-08-19 1a9a052
[liblsl] Fix resolver recursion issue
