### 2026-08-20 8253092
Added more LSL timing info now that liblsl package has it
