+++ packages/liblsl/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.13.3+0
+
+- Updated build hook to add the flag `-Wl,--allow-shlib-undefined` for Android builds, to prevent an error when building the shared library.
+
 # 0.13.2+0
 
 - Zenodo DOI: [10.5281/zenodo.20340248](https://doi.org/10.5281/zenodo.20340248)
