### 2026-09-14 6202ea1
[peer_coodinator] add inlet op locking (for async ops)
