### 2026-09-14 a81e25a
[liblsl_coordinator] Async + out of loop inlet adding
