+++ packages/liblsl_coordinator/pubspec.yaml
@@ -21,6 +21,9 @@ dependencies:
   # liblsl: ^0.9.0
   liblsl:
     path: ../liblsl
+  # Direct, not transitive: the inlet worker allocates an out-parameter for
+  # lsl_open_stream when it opens inlets on a helper isolate.
+  ffi: ^2.2.0
   synchronized: ^3.4.0+1
   logging: ^1.3.0
   uuid: ^4.5.3
