+++ packages/peer_coordinator/WEBRTC_PLAN.md
@@ -1,330 +1,0 @@
-# A genuinely peer-to-peer transport, over WebRTC data channels
-
-Originally a handoff plan written at the end of the session that built phase 1.
-**Phases 1–6 are now implemented**; see Status below for what is verified and
-what still needs a device. The design sections below are kept as the record of
-why the transport is shaped the way it is — read `packages/webrtc_coordinator/README.md`
-for how to use it.
-
-## Why
-
-The WebSocket transport relays every byte. The hub is on the data path by
-design — `WsSampleFrame`'s header says so ("srcSlot REWRITTEN BY THE HUB on
-relay"), and `addInlet` does not connect to anything, it installs a route in the
-hub. Every message and every sample costs two network hops.
-
-The question this answers: can the hub be demoted to discovery and connection
-setup, with data going directly between peers? Yes. The abstraction was built
-for it — LSL is already genuinely peer-to-peer behind the same interfaces — and
-**no core changes are needed**.
-
-WebRTC rather than a listening WebSocket per peer, because browsers cannot
-listen. A direct-WS mesh would drop the web target that `peer_coordinator`
-currently supports; WebRTC keeps it.
-
-## Status
-
-### Done — phase 1: the signalling frame
-
-`WsControl.signal` is the hub's only unicast and the only frame whose contents it
-never inspects. 188 package tests pass, analyzer is clean, and the web-safety
-gate still compiles.
-
-- `lib/src/websocket/ws_protocol.dart` — `WsControl.signal`; `wsProtocolVersion`
-  bumped 1 → 2. (It is 3 now: `WsControl.unsubscribe` followed, for
-  `removeInlet`.)
-- `lib/hub.dart` — `CoordinationHub._forwardSignal`. Verifies the sender owns the
-  `from` endpoint (exactly as `_relayMessage` does), then re-encodes rather than
-  forwarding bytes, so `from` is the endpoint the hub verified and not whatever
-  the sender wrote. Unknown or departed `to` is dropped silently.
-- `lib/src/websocket/ws_connection.dart` — `sendSignal(...)`, `Stream<WsSignal>
-  signals`, and the `WsSignal` class. `_signals` is closed in `close()`.
-- `lib/websocket.dart` — exports `WsSignal`.
-- `test/websocket/ws_signalling_test.dart` — 6 tests: delivery, opaque payloads,
-  unicast isolation, impersonation refusal, unknown target, bidirectional
-  offer/answer.
-- `test/websocket/ws_protocol_test.dart` — signal round-trip, and the version
-  coupling pinned out loud.
-
-**The version bump is a hard cut.** `WsFrame.decode` rejects unknown versions
-rather than guessing, so a hub and its clients must be deployed together.
-
-### Done — phases 2–6
-
-All in `packages/webrtc_coordinator` (pure Dart) and
-`packages/webrtc_coordinator_flutter` (the plugin binding).
-
-- **Phase 2 — seam, config, transport.** `lib/src/rtc/rtc_adapter.dart` is the
-  boundary; `lib/src/rtc/fake/fake_adapter.dart` is the pure-Dart fake behind it.
-  `RtcTransportConfig` / `RtcTransport` in `lib/src/transport/rtc_transport.dart`,
-  with discovery delegated to `WsDiscovery` verbatim and a non-null
-  `PeerClockOffsets`. The adapter is supplied by an `RtcAdapterFactory` on the
-  config — the one seam that keeps `flutter_webrtc` out of the pure-Dart package.
-- **Phase 3 — the data path.** `RtcMesh` (`lib/src/rtc/rtc_mesh.dart`) owns links
-  and channels; `RtcStreamMixin` + `RtcCoordinationStream` / `RtcDataStream` /
[…]
