### 2026-09-16 af1507b
Fix resolver leak + formatting
