+++ packages/liblsl_analysis/pubspec.yaml
@@ -39,7 +39,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.9
-  file_picker: ^12.0.0-beta.5
+  file_picker: ^13.1.0
   dartframe:
     git:
       url: https://github.com/NexusDynamic/DartFrame
