+++ packages/liblsl_timing/pubspec.yaml
@@ -46,7 +46,7 @@ dependencies:
   csv: ^8.0.0
   path: ^1.9.1
   shared_preferences: ^2.5.5
-  file_picker: ^12.0.0-beta.5
+  file_picker: ^13.1.0
   synchronized: ^3.4.0+1
   easy_localization: ^3.0.8
   flutter_fullscreen: ^1.2.0
