+++ packages/liblsl/pubspec.yaml
@@ -14,7 +14,7 @@ environment:
 dependencies:
   ffi: ^2.2.0
   hooks: ^2.2.0
-  native_toolchain_c: ^0.19.4
+  native_toolchain_c: ^0.19.5 # pinned to 0.19.4 because 0.19.5 breaks building on linux arm right now.
   code_assets: ^1.2.1
   logging: ^1.3.0
   meta:
@@ -22,7 +22,7 @@ dependencies:
   android_libcpp_shared: ^0.3.0
 
 dev_dependencies:
-  ffigen: ^21.0.0
+  ffigen: ^22.0.0
   lints: ^6.1.0
   test: ^1.31.0
   dartdoc: ^9.0.8
