+++ packages/liblsl/pubspec.yaml
@@ -14,7 +14,7 @@ environment:
 dependencies:
   ffi: ^2.2.0
   hooks: ^2.2.0
-  native_toolchain_c: ^0.19.5 # pinned to 0.19.4 because 0.19.5 breaks building on linux arm right now.
+  native_toolchain_c: ^0.19.4 # pinned to 0.19.4 because 0.19.5 breaks building on linux arm right now.
   code_assets: ^1.2.1
   logging: ^1.3.0
   meta:
@@ -29,6 +29,7 @@ dev_dependencies:
 
 dependency_overrides:
   code_assets: ^1.0.0 # temporarily override to avoid issues with equality, breaking change with code_assets: ^2.0.0
+  native_toolchain_c: 0.19.4 # pinned to 0.19.4 because 0.19.5 breaks building on linux arm right now.
 
 ffigen:
   name: Liblsl
