### 2026-09-21 cf813f5
Update pubspec, remove override for code_assets
