+++ packages/liblsl/pubspec.yaml
@@ -15,7 +15,7 @@ dependencies:
   ffi: ^2.2.0
   hooks: ^2.2.0
   native_toolchain_c: ^0.19.4 # pinned to 0.19.4 because 0.19.5 breaks building on linux arm right now.
-  code_assets: ^1.2.1
+  code_assets: ^2.1.0
   logging: ^1.3.0
   meta:
   fast_immutable_collections: ^11.1.0
