### 2026-09-24 2e6561e
Add signal_core signal engine
