+++ packages/signal_core/pubspec.yaml
@@ -1,0 +1,22 @@
+name: signal_core
+description: >-
+  Format-agnostic multichannel signal engine: summary pyramids for fast
+  zooming, display filters (SOS, zero-phase and streaming), re-referencing,
+  channel quality and random-access byte sources. Pure Dart, web-safe.
+version: 0.1.0
+homepage: https://zeyus.com/
+repository: https://github.com/NexusDynamic/liblsl.dart
+issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
+publish_to: 'none'
+
+resolution: workspace
+
+environment:
+  sdk: ^3.10.0
+
+dependencies:
+  web: ^1.1.0
+
+dev_dependencies:
+  lints: ^6.0.0
+  test: ^1.28.0
