### 2026-09-24 ab7e9ae
Add signal_viewer and signal_viewer_lsl
