+++ packages/signal_viewer/CHANGELOG.md
@@ -1,0 +1,4 @@
+## 0.1.0
+
+- Initial version, extracted from hyprview 0.5.1: the viewer (tabs, plots,
+  controls, dialogs, preferences) with `SourceProvider`s for what it opens.
