+++ packages/signal_viewer_lsl/pubspec.yaml
@@ -1,0 +1,32 @@
+name: signal_viewer_lsl
+description: >-
+  Lab Streaming Layer source provider for signal_viewer: discover, view and
+  publish LSL streams (desktop and mobile; a no-op on the web).
+version: 0.1.0
+homepage: https://zeyus.com/
+repository: https://github.com/NexusDynamic/liblsl.dart
+issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
+publish_to: 'none'
+
+resolution: workspace
+
+environment:
+  sdk: ^3.10.0
+
+dependencies:
+  flutter:
+    sdk: flutter
+  # Local-network permission (iOS, Android) and the Android multicast lock
+  # that stream discovery needs.
+  flutter_multicast_lock: ^1.2.1
+  liblsl: any
+  permission_handler: ^13.0.1
+  signal_viewer: ^0.1.0
+
+dev_dependencies:
+  flutter_lints: ^6.0.0
+  flutter_test:
+    sdk: flutter
+
+flutter:
+  uses-material-design: true
