+++ packages/xdf/dart_test.yaml
@@ -1,0 +1,5 @@
+override_platforms:
+  chrome:
+    settings:
+      executable:
+        linux: chromium
