### 2026-09-24 7df59d2
Add signal_viewer_xdf for parsing xdf files in the viewer
