+++ pubspec.yaml
@@ -18,6 +18,7 @@ workspace:
   - packages/signal_core
   - packages/signal_viewer
   - packages/signal_viewer_lsl
+  - packages/signal_viewer_xdf
   - packages/xdf
 dev_dependencies:
   melos: ^8.2.2
