### 2026-09-24 fa15056
LSL stream info metadata, clock offset, measured rate
