+++ packages/signal_viewer_lsl/pubspec.yaml
@@ -22,6 +22,7 @@ dependencies:
   liblsl: any
   permission_handler: ^13.0.1
   signal_viewer: ^0.1.0
+  xml: ^6.5.0
 
 dev_dependencies:
   flutter_lints: ^6.0.0
