### 2026-09-25 4076af6
Add serial_transport for webserial and desktop
