+++ packages/serial_transport/src/serial_shim.c
@@ -1,0 +1,594 @@
+/*
+    Minimal cross-platform serial port access for serial_transport.
+    See serial_shim.h.
+
+    Copyright (c) 2025 zeyus. MIT License (see LICENSE).
+
+    Port configuration follows stream_serial.cpp of hyperscanner_lib
+    (Aarhus University).
+*/
+
+#if !defined(_WIN32)
+#define _DEFAULT_SOURCE /* cfmakeraw, realpath, strncasecmp on glibc */
+#endif
+
+#include "serial_shim.h"
+
+#include <stdarg.h>
+#include <stdio.h>
+#include <string.h>
+
+/* ------------------------------------------------------------------------ */
+/* Shared helpers                                                           */
+/* ------------------------------------------------------------------------ */
+
+static void set_err(char *err, int err_len, const char *fmt, ...)
+{
+    if (err == NULL || err_len <= 0)
+        return;
+    va_list args;
+    va_start(args, fmt);
+    vsnprintf(err, (size_t)err_len, fmt, args);
+    va_end(args);
+}
+
+/* Appends to a caller buffer, tracking the size needed for everything. */
+typedef struct
+{
+    char *buf;
+    int cap;
+    int needed; /* excluding the terminating NUL */
+} out_buf_t;
+
+static void out_append(out_buf_t *out, const char *s)
+{
+    size_t n = strlen(s);
+    if (out->buf != NULL && out->cap > 0)
+    {
+        int room = out->cap - 1 - out->needed;
+        if (room > 0)
+        {
+            size_t copy = n < (size_t)room ? n : (size_t)room;
+            memcpy(out->buf + out->needed, s, copy);
+            out->buf[out->needed + (int)copy] = '\0';
+        }
+    }
+    out->needed += (int)n;
+}
+
+/* Tabs and newlines separate fields and records, so keep them out of values. */
+static void sanitize(char *s)
+{
+    for (; *s; ++s)
+        if (*s == '\t' || *s == '\n' || *s == '\r')
+            *s = ' ';
+}
+
+static void out_port(out_buf_t *out, const char *path, char *desc,
+                     const char *vid, const char *pid, char *product)
+{
+    sanitize(desc);
+    sanitize(product);
+    out_append(out, path);
+    out_append(out, "\t");
+    out_append(out, desc);
+    out_append(out, "\t");
+    out_append(out, vid);
+    out_append(out, "\t");
+    out_append(out, pid);
+    out_append(out, "\t");
+    out_append(out, product);
+    out_append(out, "\n");
+}
+
+#if defined(_WIN32)
+/* ======================================================================== */
+/* Windows                                                                  */
+/* ======================================================================== */
+
+#include <windows.h>
+#include <setupapi.h>
+#include <devpropdef.h>
+#include <devguid.h>
+
+/* DEVPKEY_Device_BusReportedDeviceDesc, defined here as in stream_serial.cpp */
+static const DEVPROPKEY hss_DEVPKEY_Device_BusReportedDeviceDesc = {
+    {0x540b947e, 0x8b40, 0x45bc, {0xa8, 0xa2, 0x6a, 0x0b, 0x89, 0x4c, 0xbd, 0xa2}}, 4};
+
+static void win_err(char *err, int err_len, const char *what)
+{
+    DWORD code = GetLastError();
+    char msg[256] = "";
[…]
