### 2026-09-25 2b06d8b
serial_transport select baud rate
