### 2026-09-25 a81586c
serial_transport fix read hanging in flutter
