### 2026-09-25 55d4fde
Add signal_viewer_serial
