+++ packages/signal_viewer_serial/CHANGELOG.md
@@ -1,0 +1,4 @@
+## 0.1.0
+
+- Initial version: `SerialStreamProvider` and `SerialStreamSession` read
+  lines of numbers (or name:value pairs) from serial ports as live streams.
