+++ packages/signal_viewer_serial/pubspec.yaml
@@ -1,0 +1,25 @@
+name: signal_viewer_serial
+description: >-
+  Serial devices for signal_viewer: lines of numbers from serial ports
+  (e.g. an Arduino) as live streams, on desktop and the web (WebSerial).
+version: 0.1.0
+homepage: https://zeyus.com/
+repository: https://github.com/NexusDynamic/liblsl.dart
+issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
+publish_to: 'none'
+
+resolution: workspace
+
+environment:
+  sdk: ^3.10.0
+
+dependencies:
+  flutter:
+    sdk: flutter
+  serial_transport: ^0.1.0
+  signal_viewer: ^0.1.0
+
+dev_dependencies:
+  flutter_lints: ^6.0.0
+  flutter_test:
+    sdk: flutter
