+++ pubspec.yaml
@@ -21,6 +21,7 @@ workspace:
   - packages/signal_viewer_xdf
   - packages/xdf
   - packages/serial_transport
+  - packages/signal_viewer_serial
   - apps/lsl_viewer
 dev_dependencies:
   melos: ^8.2.2
