+++ packages/lsl_tools/pubspec.yaml
@@ -1,0 +1,30 @@
+name: lsl_tools
+description: >-
+  Lab Streaming Layer tools in pure Dart: a web-safe facade over liblsl,
+  recording to XDF, a WebSocket bridge between networks, test outlets, and
+  the `lsl` command line tool.
+version: 0.1.0
+homepage: https://zeyus.com/
+repository: https://github.com/NexusDynamic/liblsl.dart
+issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
+publish_to: 'none'
+
+resolution: workspace
+
+environment:
+  sdk: ^3.10.0
+
+dependencies:
+  args: ^2.6.0
+  liblsl: any
+  signal_core: ^0.1.0
+  web_socket_channel: ^3.0.3
+  xdf: ^0.1.0
+  xml: ^6.5.0
+
+dev_dependencies:
+  lints: ^6.0.0
+  test: ^1.28.0
+
+executables:
+  lsl:
