+++ packages/openbci_cyton/pubspec.yaml
@@ -1,7 +1,7 @@
-name: signal_viewer_serial
+name: openbci_cyton
 description: >-
-  Serial devices for signal_viewer: lines of numbers from serial ports
-  (e.g. an Arduino) as live streams, on desktop and the web (WebSerial).
+  The OpenBCI Cyton (and Cyton + Daisy) biosensing board over its USB
+  dongle, in pure Dart: commands, channel settings and samples in µV.
 version: 0.1.0
 homepage: https://zeyus.com/
 repository: https://github.com/NexusDynamic/liblsl.dart
@@ -14,13 +14,8 @@ environment:
   sdk: ^3.10.0
 
 dependencies:
-  flutter:
-    sdk: flutter
-  openbci_cyton: ^0.1.0
   serial_transport: ^0.1.0
-  signal_viewer: ^0.1.0
 
 dev_dependencies:
-  flutter_lints: ^6.0.0
-  flutter_test:
-    sdk: flutter
+  lints: ^6.0.0
+  test: ^1.28.0
