### 2026-09-25 e8c841b
Serial devices for android as well
