+++ apps/lsl_viewer/README.md
@@ -10,9 +10,16 @@ It is `signal_viewer` with two source providers:
   each in a tab.
 - `XdfProvider` (`signal_viewer_xdf`): open .xdf files of any size, a tab
   per stream.
-- `SerialStreamProvider` (`signal_viewer_serial`): serial devices that
-  print lines of numbers (e.g. an Arduino), on desktop and in browsers
-  with WebSerial.
+- `SerialStreamProvider` (`signal_viewer_serial`): the OpenBCI Cyton (+
+  Daisy), and serial devices that print lines of numbers (e.g. an
+  Arduino): on desktop, on Android over USB, and in Chrome and Edge with
+  WebSerial.
+
+In a browser, LSL itself is not available (liblsl needs the system's
+sockets and threads): serial devices and XDF files work, and live LSL
+streams can be viewed through a bridge (LSL > Connect to an LSL bridge…,
+shared from a desktop with "Share streams over the network…" or
+`lsl share`).
 
 Each tab has filters, re-referencing, power, signal quality and trigger
 decoding; the source setup (View menu) overrides a stream's type and
