### 2026-09-25 3a39922
Tag-driven per-package releases; liblsl 1.0.0
