+++ packages/liblsl/src/wasm-demo/build-liblsl-emcc.sh
@@ -1,50 +1,0 @@
-LSL_SRC=./src/liblsl-9e3823bb
-emcc -sEXPORT_NAME=liblsl --no-entry \
-      -sENVIRONMENT=web,worker,node -DVERSION=1.16.2 -DLSL_ABI_VERSION=2 \
-      -DLOGURU_PTHREADS=0 -DLOGURU_WITH_STREAMS=0 \
-      -DASIO_DISABLE_THREADS -D_POSIX_C_SOURCE=200809L \
-      --std=c++17 -sERROR_ON_UNDEFINED_SYMBOLS=0 -sALLOW_MEMORY_GROWTH=1 \
-      -sNO_DISABLE_EXCEPTION_CATCHING \
-      --js-library stub_pthread.js \
-      -sEXPORTED_FUNCTIONS=_lsl_last_error,_lsl_protocol_version,_lsl_library_version,_lsl_library_info,_lsl_local_clock,_lsl_destroy_string,_lsl_set_config_filename,_lsl_set_config_content,_lsl_create_inlet,_lsl_create_inlet_ex,_lsl_destroy_inlet,_lsl_get_fullinfo,_lsl_open_stream,_lsl_close_stream,_lsl_time_correction,_lsl_time_correction_ex,_lsl_set_postprocessing,_lsl_pull_sample_f,_lsl_pull_sample_d,_lsl_pull_sample_l,_lsl_pull_sample_i,_lsl_pull_sample_s,_lsl_pull_sample_c,_lsl_pull_sample_str,_lsl_pull_sample_buf,_lsl_pull_sample_v,_lsl_pull_chunk_f,_lsl_pull_chunk_d,_lsl_pull_chunk_l,_lsl_pull_chunk_i,_lsl_pull_chunk_s,_lsl_pull_chunk_c,_lsl_pull_chunk_str,_lsl_pull_chunk_buf,_lsl_samples_available,_lsl_inlet_flush,_lsl_was_clock_reset,_lsl_smoothing_halftime,_lsl_create_outlet,_lsl_create_outlet_ex,_lsl_destroy_outlet,_lsl_push_sample_f,_lsl_push_sample_d,_lsl_push_sample_l,_lsl_push_sample_i,_lsl_push_sample_s,_lsl_push_sample_c,_lsl_push_sample_str,_lsl_push_sample_v,_lsl_push_sample_ft,_lsl_push_sample_dt,_lsl_push_sample_lt,_lsl_push_sample_it,_lsl_push_sample_st,_lsl_push_sample_ct,_lsl_push_sample_strt,_lsl_push_sample_vt,_lsl_push_sample_ftp,_lsl_push_sample_dtp,_lsl_push_sample_ltp,_lsl_push_sample_itp,_lsl_push_sample_stp,_lsl_push_sample_ctp,_lsl_push_sample_strtp,_lsl_push_sample_vtp,_lsl_push_sample_buf,_lsl_push_sample_buft,_lsl_push_sample_buftp,_lsl_push_chunk_f,_lsl_push_chunk_d,_lsl_push_chunk_l,_lsl_push_chunk_i,_lsl_push_chunk_s,_lsl_push_chunk_c,_lsl_push_chunk_str,_lsl_push_chunk_ft,_lsl_push_chunk_dt,_lsl_push_chunk_lt,_lsl_push_chunk_it,_lsl_push_chunk_st,_lsl_push_chunk_ct,_lsl_push_chunk_strt,_lsl_push_chunk_ftp,_lsl_push_chunk_dtp,_lsl_push_chunk_ltp,_lsl_push_chunk_itp,_lsl_push_chunk_stp,_lsl_push_chunk_ctp,
[…]
