+++ packages/liblsl/src/wasm-demo/stub_pthread.js
@@ -1,6 +1,0 @@
-// Stub library to provide missing pthread functions
-mergeInto(LibraryManager.library, {
-  pthread_getname_np: function() {
-    return 0; // Return success
-  }
-});
