### 2026-09-25 f9997ae
CI: run LSL tests on every PR and release
