### 2026-09-25 7784bef
Getting signal_core and xdf ready for publishing
