+++ packages/xdf/pubspec.yaml
@@ -8,7 +8,6 @@ version: 0.1.0
 homepage: https://zeyus.com/
 repository: https://github.com/NexusDynamic/liblsl.dart
 issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
-publish_to: 'none'
 
 resolution: workspace
 
