### 2026-09-25 4ac5c66
Tests: fix open-file exhaustion and a native-assets race
