### 2026-09-25 e2ca882
liblsl: raise the soft open-file limit on load (macOS, Linux)
