### 2026-09-25 0b2f10f
Prepare coordinator packages for pub.dev
