+++ packages/webrtc_coordinator_flutter/CHANGELOG.md
@@ -1,0 +1,4 @@
+## 0.1.0
+
+- Initial release: `flutterWebrtcAdapterFactory`, an `RtcPeerAdapter` for
+  `webrtc_coordinator` backed by real `flutter_webrtc` peer connections.
