+++ packages/liblsl_coordinator/pubspec.yaml
@@ -4,9 +4,6 @@ version: 0.4.1
 homepage: https://zeyus.com/
 repository: https://github.com/NexusDynamic/liblsl.dart
 issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
-# for now, we aren't publishing, it needs to use an available LSL version
-# not from path.
-publish_to: 'none'
 
 resolution: workspace
 
@@ -15,12 +12,9 @@ environment:
 
 # Add regular dependencies here.
 dependencies:
-  meta:
-  peer_coordinator:
-    path: ../peer_coordinator
-  # liblsl: ^0.9.0
-  liblsl:
-    path: ../liblsl
+  meta: ^1.15.0
+  peer_coordinator: ^0.3.1
+  liblsl: ^1.0.0
   # Direct, not transitive: the inlet worker allocates an out-parameter for
   # lsl_open_stream when it opens inlets on a helper isolate.
   ffi: ^2.2.0
