+++ packages/webrtc_coordinator_flutter/pubspec.yaml
@@ -7,7 +7,6 @@ version: 0.1.0
 homepage: https://zeyus.com/
 repository: https://github.com/NexusDynamic/liblsl.dart
 issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
-publish_to: 'none'
 
 resolution: workspace
 
@@ -20,10 +19,8 @@ dependencies:
     sdk: flutter
   flutter_webrtc: ^1.6.0
   logging: ^1.3.0
-  webrtc_coordinator:
-    path: ../webrtc_coordinator
-  peer_coordinator:
-    path: ../peer_coordinator
+  webrtc_coordinator: ^0.1.0
+  peer_coordinator: ^0.3.1
 
 dev_dependencies:
   flutter_test:
