+++ packages/webrtc_coordinator/pubspec.yaml
@@ -7,7 +7,6 @@ version: 0.1.0
 homepage: https://zeyus.com/
 repository: https://github.com/NexusDynamic/liblsl.dart
 issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
-publish_to: 'none'
 
 resolution: workspace
 
@@ -15,12 +14,11 @@ environment:
   sdk: ^3.12.2
 
 dependencies:
-  meta:
+  meta: ^1.15.0
   logging: ^1.3.0
   collection: ^1.19.1
   fast_immutable_collections: ^11.2.0
-  peer_coordinator:
-    path: ../peer_coordinator
+  peer_coordinator: ^0.3.1
   # Regular dep, not dev: lib/testing.dart ships the fake adapter and the
   # participation harness for webrtc_coordinator_flutter to reuse, exactly as
   # peer_coordinator ships its conformance suite.
