+++ packages/peer_coordinator/pubspec.yaml
@@ -6,8 +6,6 @@ version: 0.3.1
 homepage: https://zeyus.com/
 repository: https://github.com/NexusDynamic/liblsl.dart
 issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
-# Not published yet; versioned alongside liblsl_coordinator for now.
-publish_to: 'none'
 
 resolution: workspace
 
@@ -15,7 +13,7 @@ environment:
   sdk: ^3.12.2
 
 dependencies:
-  meta: any
+  meta: ^1.15.0
   # HMAC-SHA256 for the hub auth handshake. Pure Dart, so it stays web-safe.
   crypto: ^3.0.6
   synchronized: ^3.4.0+1
