### 2026-09-25 533578e
melos: merge test stderr into stdout so liblsl log lines are not labelled ERROR
