### 2026-09-25 739ae02
updated some readme / info
