+++ packages/serial_transport/README.md
@@ -1,0 +1,3 @@
+# serial transport
+
+Multi-platform serial transport shim for Dart.
