### 2026-09-25 fdcaa5d
add liblsl.dart benchmarks (customSmallerIsBetter) benchmark result for 739ae0200b217d4995327111fdf7e8b0a6f081bf
