### 2026-09-25 893173f
[signal_viewer_lsl] bridge relay/replay
