+++ packages/signal_viewer_lsl/pubspec.yaml
@@ -28,6 +28,7 @@ dev_dependencies:
   flutter_lints: ^6.0.0
   flutter_test:
     sdk: flutter
+  signal_viewer_xdf: ^0.1.0
 
 flutter:
   uses-material-design: true
