### 2026-09-25 1919aa1
[ci] lsl command line tools bundled with app
