### 2026-09-25 ddc6fa4
[lsl_viewer] placeholder app icon
