+++ apps/lsl_viewer/linux/CMakeLists.txt
@@ -94,6 +94,11 @@ set(INSTALL_BUNDLE_LIB_DIR "${CMAKE_INSTALL_PREFIX}/lib")
 install(TARGETS ${BINARY_NAME} RUNTIME DESTINATION "${CMAKE_INSTALL_PREFIX}"
   COMPONENT Runtime)
 
+# The window icon (see runner/my_application.cc) and the .desktop entry.
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/../assets/icon/icon.png"
+  "${CMAKE_CURRENT_SOURCE_DIR}/lsl_viewer.desktop"
+  DESTINATION "${INSTALL_BUNDLE_DATA_DIR}" COMPONENT Runtime)
+
 install(FILES "${FLUTTER_ICU_DATA_FILE}" DESTINATION "${INSTALL_BUNDLE_DATA_DIR}"
   COMPONENT Runtime)
 
