### 2026-09-25 b3ef0ce
[lsl_tools] tool setup guide
