+++ packages/lsl_tools/doc/relay.md
@@ -1,0 +1,318 @@
+# Sharing LSL streams over the network: bridge and relay
+
+Lab Streaming Layer finds streams with multicast, so it only works where
+multicast reaches: usually one local network. A web browser can't use LSL
+at all, because it has no raw sockets. The **LSL bridge** in `lsl_tools` carries
+streams over a WebSocket instead:
+
+- **Across networks:** another building, a VPN, a cloud server.
+- **To browsers:** the [web version of LSL Viewer](https://nexusdynamic.org/liblsl.dart/lsl_viewer/)
+  can view live streams and publish its own.
+- **Between browsers:** as a relay, what one client publishes the others
+  receive, e.g. an XDF recording replayed in one browser and viewed in
+  another.
+
+You can run it from the **LSL Viewer** desktop app (no command line needed)
+or with the **`lsl` command line tool**.
+
+## Contents
+
+- [Which setup do I need?](#which-setup-do-i-need)
+- [Getting the tools](#getting-the-tools)
+- [Quick start: view lab streams in a browser](#quick-start-view-lab-streams-in-a-browser)
+- [Setup 1: streams from one LSL network to another](#setup-1-streams-from-one-lsl-network-to-another)
+- [Setup 2: a browser publishes into LSL](#setup-2-a-browser-publishes-into-lsl)
+- [Setup 3: a relay between browsers](#setup-3-a-relay-between-browsers)
+- [Security: tokens, origins and TLS](#security-tokens-origins-and-tls)
+- [Running a relay on a server (wss://)](#running-a-relay-on-a-server-wss)
+- [Command reference](#command-reference)
+- [Limitations](#limitations)
+- [Troubleshooting](#troubleshooting)
+
+## Which setup do I need?
+
+| You want to… | Run this where the streams are | Clients connect with |
+| --- | --- | --- |
+| See lab streams in a browser | `lsl share` (or viewer: *Share*) | Viewer: *LSL > Connect to an LSL bridge…* |
+| Get lab streams into LSL on another network | `lsl share` | `lsl bridge ws://host:8765` (or viewer: *Publish its streams on LSL here*) |
+| Get a browser's streams (serial device, replayed XDF) into LSL | `lsl share --accept` (or viewer: *Share + accept*) | Viewer in the browser: *Forward…* or *Replay…* |
+| Pass streams between browsers, no LSL anywhere | `lsl relay` (or viewer: *Relay only*) | Viewers: *Connect*, then *Forward*/*Replay* and *View* |
+| Send lab streams out to a relay on a server (the lab accepts no incoming connections) | `lsl publish wss://relay…` in the lab | Anyone connected to the relay |
+
+```text
+Setup 1   [LSL network A] ──lsl share──▶ ws ──▶ lsl bridge ──▶ [LSL network B]
+                                          └───▶ browser (LSL Viewer on the web)
+
+Setup 2   browser ──publish──▶ ws ──▶ lsl share --accept ──▶ [LSL on that computer]
+                                                        └──▶ other clients
+
+Setup 3   browser A ──publish──▶ ws ──▶ lsl relay ──▶ ws ──▶ browser B, C, lsl bridge…
+          [LSL lab] ──lsl publish──▶ ┘
+```
+
+Streams that clients publish are always shared with every other client. On
+a *share* or *share + accept* server they also become ordinary LSL streams
+on that computer. A *relay* only passes them between clients, so it doesn't
+need LSL.
+
+## Getting the tools
+
+- **LSL Viewer (desktop):** download from the
+  [releases page](https://github.com/NexusDynamic/liblsl.dart/releases)
+  (`lsl_viewer <version>`). The sharing controls are under
+  *LSL > Share or relay streams…*.
+- **`lsl` command line tool:** each LSL Viewer release also has
[…]
