### 2026-09-25 d800305
[docs] landing site, relay guide on Pages, READMEs
