+++ packages/lsl_tools/README.md
@@ -1,30 +1,81 @@
 # lsl_tools
 
-Lab Streaming Layer tools in pure Dart, and the `lsl` command line tool.
+Lab Streaming Layer tools in pure Dart, and the `lsl` command line tool:
+list, record to XDF, share across networks, bridge, relay, replay and
+generate LSL streams.
 
-- `lsl`: a facade over `package:liblsl` that is safe to import on the web
-  (where `lsl.supported` is false): discovery, inlets, outlets.
-- `LslRecorder`: record streams to XDF as LabRecorder does (raw time
+- **`lsl`:** a facade over `package:liblsl` that is safe to import on the
+  web, where `lsl.supported` is false. It covers discovery, inlets and
+  outlets.
+- **`LslRecorder`:** records streams to XDF as LabRecorder does (raw time
   stamps, clock offsets, footers).
-- `LslBridgeServer` / `LslBridgeClient`: share streams over a WebSocket,
-  across networks multicast does not reach (the client also runs in
-  browsers).
-- `LslSignalGenerator`, `LslMarkerSender`: test outlets.
+- **`LslBridgeServer`, `LslBridgeClient`, `LslBridgeRepublisher`:** carry
+  LSL streams over a WebSocket, across networks multicast doesn't reach and
+  to and from browsers. The server can also relay streams between its
+  clients. The client runs in browsers too.
+- **`LslSignalGenerator`, `LslMarkerSender`:** test outlets.
 
-## lsl
+**Setting up a bridge or relay:** see [doc/relay.md](doc/relay.md), also
+published at <https://nexusdynamic.org/liblsl.dart/relay.html>.
+
+## The `lsl` command line tool
+
+Prebuilt bundles are attached to each
+[`lsl_viewer` release](https://github.com/NexusDynamic/liblsl.dart/releases?q=lsl_viewer&expanded=true)
+as `lsl-cli-<version>-<os>`. Run `bin/lsl` and keep `lib/`, which holds
+liblsl, next to it. From a clone, use `dart run lsl_tools:lsl …`, or
+`dart build cli` for the same bundle.
 
 ```sh
-dart run lsl_tools:lsl list
-dart run lsl_tools:lsl info EEG
-dart run lsl_tools:lsl record session.xdf -s EEG -s Markers   # Ctrl-C stops
-dart run lsl_tools:lsl share --port 8765 --token secret      # on the lab network
-dart run lsl_tools:lsl bridge ws://lab-pc:8765 --token secret # elsewhere
-dart run lsl_tools:lsl share --accept                        # take streams
-                                        # clients publish (e.g. a browser)
-dart run lsl_tools:lsl replay session.xdf --loop
-dart run lsl_tools:lsl generate -c 8 -r 500
+lsl list [--wait 2]                              # streams on the network
+lsl info EEG                                     # full XML and clock offset
+lsl record session.xdf -s EEG -s Markers [-d 60] # Ctrl-C stops
+lsl replay session.xdf [--loop]                  # an XDF file as LSL streams
+lsl generate -c 8 -r 500 [-f 10] [--name --type] # a test stream
+lsl cyton [/dev/ttyUSB0]                         # OpenBCI Cyton (+ Daisy)
+
+# Across networks and to browsers (see doc/relay.md)
+lsl share [-s EEG] [--token t] [-p 8765] [--host 0.0.0.0]
+          [--accept [--no-local-outlets]] [--rescan 5] [--allow-origin https://…]
+lsl relay [--token t] [-p 8765] [--host 127.0.0.1] [--allow-origin …]
+lsl bridge ws://host:8765 [-s EEG] [--token t] [--suffix " (remote)"]
+lsl publish wss://relay.example.org [-s EEG] [--token t] [--rescan 5]
 ```
 
[…]
