+++ packages/signal_core/pubspec.yaml
@@ -7,6 +7,12 @@ version: 0.1.0
 homepage: https://zeyus.com/
 repository: https://github.com/NexusDynamic/liblsl.dart
 issue_tracker: https://github.com/NexusDynamic/liblsl.dart/issues
+topics:
+  - signal-processing
+  - eeg
+  - neuroscience
+  - dsp
+  - visualization
 
 resolution: workspace
 
