+++ robots.txt
@@ -1,0 +1,3 @@
+User-agent: *
+Allow: /
+Sitemap: https://nexusdynamic.org/liblsl.dart/sitemap.xml
