### 2026-09-25 bc3ab79
add liblsl.dart benchmarks (customSmallerIsBetter) benchmark result for 6aedf52e1dbd9896aef80341135a4af228db6cae
