### 2026-09-25 3fa2153
[lsl_viewer] version bump
