+++ apps/lsl_viewer/pubspec.yaml
@@ -1,7 +1,7 @@
 name: lsl_viewer
 description: "Lab Streaming Layer stream and XDF recording viewer"
 publish_to: 'none'
-version: 0.1.0+1
+version: 0.1.1+1
 
 resolution: workspace
 
