+++ lsl_viewer/flutter_bootstrap.js
@@ -42,6 +42,6 @@ _flutter.buildConfig = {"engineRevision":"af7e796e161ae0bb1ff0758c71a7105418bd9d
 
 _flutter.loader.load({
   serviceWorkerSettings: {
-    serviceWorkerVersion: "96068072" /* Flutter's service worker is deprecated and will be removed in a future Flutter release. */
+    serviceWorkerVersion: "4153028879" /* Flutter's service worker is deprecated and will be removed in a future Flutter release. */
   }
 });
